### 2025-10-31 #9332 @editorialbot
[PRE REVIEW]: Discovering the SUPER in computing - dagster-slurm for reproducible research on HPC
**Submitting author:** @HPicatto (<a href="http://orcid.org/0000-0001-6662-464X">Hernan Picatto</a>)
**Repository:** https://github.com/ascii-supply-networks/dagster-slurm
**Branch with paper.md** (empty if default branch):
**Version:** v1.7.0
