### 2025-10-31 ascii-supply-networks/dagster-slurm#76 @HPicatto
publishing with JOSS
This issue will follow the state of our JOSS submission
currently in [prereview](https://github.com/openjournals/joss-reviews/issues/9332#issuecomment-3473832885)
a version of this has also been submitted to [CHIIR26](https://chiir2026.github.io/)
@geoHeil @MaximilianHess
