### 2025-10-31 19ad4e7
fix dois (#75)
Co-authored-by: CI Hotfix <cihotfix@ascii.ac.at>
