+++ docs/paper.bib
@@ -78,7 +78,8 @@
       eprint={2408.09869},
       archivePrefix={arXiv},
       primaryClass={cs.CL},
-      url={https://arxiv.org/abs/2408.09869}, 
+      url={https://arxiv.org/abs/2408.09869},
+      doi={https://doi.org/10.48550/arXiv.2408.09869} 
 }
 
 @software{dagster,
@@ -99,7 +100,7 @@
   pages = {44--60},
   year = {2003},
   publisher = {Springer},
-  doi = {10.1007/10968987\_3}
+  doi = {https://doi.org/10.1007/10968987_3}
 }
 
 @software{pixi,
@@ -140,7 +141,7 @@ abstract = {Reproducibility is widely acknowledged as a fundamental principle in
   number={3},
   pages={77-80},
   keywords={High performance computing;Research and development;Reproducibility of results},
-  doi={10.1109/MCSE.2022.3165626}}
+  doi={https://doi.org/10.1109/MCSE.2022.3165626}}
 @misc{hayotsasson2025address,
       title={Addressing Reproducibility Challenges in HPC with Continuous Integration}, 
       author={Valérie Hayot-Sasson and Nathaniel Hudson and André Bauer and Maxime Gonthier and Ian Foster and Kyle Chard},
@@ -148,7 +149,8 @@ abstract = {Reproducibility is widely acknowledged as a fundamental principle in
       eprint={2508.21289},
       archivePrefix={arXiv},
       primaryClass={cs.DC},
-      url={https://arxiv.org/abs/2508.21289}, 
+      url={https://arxiv.org/abs/2508.21289},
+      doi={https://doi.org/10.48550/arXiv.2508.21289},
 }
 @misc{keahey_2025_15306610,
   author       = {Keahey, Kate and
@@ -175,4 +177,5 @@ abstract = {Reproducibility is widely acknowledged as a fundamental principle in
       archivePrefix={arXiv},
       primaryClass={cs.SE},
       url={https://arxiv.org/abs/2507.16166}, 
+      doi={https://doi.org/10.48550/arXiv.2507.16166},
 }
