### 2025-10-31 #9332 @danielskatz
@HPicatto - I see a few changes needed in your paper:

1. When you refer to a reference directly (by naming the authors), please remove the []s around the reference in the .md file  (See the Citations section of the [JOSS example paper](https://joss.readthedocs.io/en/latest/example_paper.html))
2. The start of your statement of need section is very confusing - it looks like two sentences have been run together, and there's an "it's" that I think should be "its".

Also, I don't really understand how this is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS.  Can you explain a bit more about what types of researchers would use this, and what types of research it would enable? And similarly, has this been cited? If not, in what situations would you expect it to be?
