### 2025-11-01 0c6f80b
feat: analytics (#77)
* feat: analytics

* feat: also pypi
