### 2025-11-01 065d9bf
1.7.0-rc.1 [skip ci]
Automatically generated by python-semantic-release
