+++ README.md
@@ -18,6 +18,8 @@ See the [docs](https://ascii-supply-networks.github.io/dagster-slurm/) for instr
 
 > **Status:** Local execution and one-job-per-asset Slurm runs are production-ready today. Session-based cluster reuse and heterogeneous jobs are actively being developed and will land in a future release.
 
+<img referrerpolicy="no-referrer-when-downgrade" src="https://telemetry.dagster-slurm.geoheil.com/a.png?x-pxid=994a20b8-4be7-4297-9f42-657b0d1f1a07&page=README.md" />
+
 ## ✨ features
 
 Hydrating supercomputers with Dagster and Slurm to enhance developer experience.
@@ -86,6 +88,3 @@ This work was kindly supported by:
 - [CSH](https://www.csh.ac.at/)
 - [EUROCC AI Hackathon](https://www.openhackathons.org/s/siteevent/a0CUP000013Tp8f2AC/se000375): This work was completed in part at the EUROCC AI Hackathon, part of the Open Hackathons program. The authors would like to acknowledge OpenACC-Standard.org for their support.
 - [Austrian Scientific Computing (ASC)](https://asc.ac.at/home/)
-
-<!-- Scarf analytics pixel -->
-<img referrerpolicy="no-referrer" src="https://static.scarf.sh/a.png?x-pxid=02e4ad6e-79c4-4ba8-b751-4833af626868&page=README.md" />
