### 2025-11-01 ec25dd8
feat: refine telemetry (#78)
* fix: use custom dns

* feat: change baseurl
