### 2025-11-01 dfef54a
feat: change logo
