### 2025-11-01 56f65cb
1.7.0-rc.2 [skip ci]
Automatically generated by python-semantic-release
