+++ .github/workflows/tutorial.yaml
@@ -40,11 +40,11 @@ jobs:
     #       --health-retries 5
     steps:
       - name: Checkout code
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
 
       - uses: prefix-dev/setup-pixi@v0.9.2
         with:
-            pixi-version: v0.58.0
+            pixi-version: v0.59.0
             cache: true
             frozen: true
       - run: cd examples && pixi run -e dev --frozen lint
