### 2025-11-01 3845b2c
feat: logo (#79)
* feat: change logo

* feat: upgrades
