+++ .github/workflows/render-pdf.yml
@@ -21,7 +21,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout repository
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
       - name: TeX and PDF
         uses: docker://openjournals/paperdraft:latest
         with:
