+++ .github/workflows/docs.yaml
@@ -31,11 +31,11 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout code
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
 
       - uses: prefix-dev/setup-pixi@v0.9.2
         with:
-            pixi-version: v0.58.0
+            pixi-version: v0.59.0
             cache: true
             frozen: true
       - run: pixi run -e docs --frozen docs-build
