+++ .github/workflows/library.yaml
@@ -59,14 +59,14 @@ jobs:
       packages: read
     steps:
       - name: Checkout code
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
 
       - name: Install uv
         uses: astral-sh/setup-uv@v6
 
       - uses: prefix-dev/setup-pixi@v0.9.2
         with:
-            pixi-version: v0.58.0
+            pixi-version: v0.59.0
             cache: true
             frozen: true
       - run: pixi run -e build --frozen lint
@@ -230,7 +230,7 @@ jobs:
           private_key: ${{ secrets.APP_PRIVATE_KEY }}
 
       - name: Checkout code
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
         with:
           fetch-depth: 0
           token: ${{ steps.generate_token.outputs.token }}
@@ -240,7 +240,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.9.2
         with:
-            pixi-version: v0.58.0
+            pixi-version: v0.59.0
             cache: true
             frozen: true
 
@@ -299,7 +299,7 @@ jobs:
           private_key: ${{ secrets.APP_PRIVATE_KEY }}
 
       - name: Checkout code
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
         with:
           fetch-depth: 0
           token: ${{ steps.generate_token.outputs.token }}
@@ -345,7 +345,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.9.2
         with:
-            pixi-version: v0.58.0
+            pixi-version: v0.59.0
             cache: true
             frozen: true
 
