### 2025-11-01 48d976f
fix: linting after upgrade (#80)
