### 2025-11-01 d92b11d
1.7.0-rc.3 [skip ci]
Automatically generated by python-semantic-release
