### 2025-11-01 de4afef
1.8.0 [skip ci]
Automatically generated by python-semantic-release
