+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <!-- version list -->
 
+## v1.8.0 (2025-11-01)
+
+
 ## v1.7.0-rc.3 (2025-11-01)
 
 ### Bug Fixes
