### 2025-11-07 #9332 @junghans
Agreed, the first sentence to the "need section" should be rewritten into two.

However, I think this integration could be useful for HPC clusters running SLURM.
