### 2025-11-10 #9332 @junghans
@MoraruMaxim could you review this?
