### 2025-11-13 e37d6fc
feat: refine slides (#81)
* feat: refine slides

* fix: address reviewers comments
