+++ docs/paper.md
@@ -39,11 +39,11 @@ aas-journal: Journal of Open Source Software
 
 # Summary
 
-Dagster is a modern data orchestrator that emphasises reproducibility, observability, and a strong developer experience [@dagster].
-In parallel, most high-performance computing (HPC) centres continue to rely on Slurm for batch scheduling and resource governance [@yoo2003slurm].
+Dagster is a modern data orchestrator that emphasises reproducibility, observability, and a strong developer experience @dagster.
+In parallel, most high-performance computing (HPC) centres continue to rely on Slurm for batch scheduling and resource governance @yoo2003slurm.
 The two ecosystems rarely meet in practice: Dagster projects often target cloud or single-node deployments, while Slurm users maintain bespoke submission scripts with limited reuse or visibility.
 This paper introduces **dagster-slurm**, an open-source integration that allows the same Dagster assets to run unchanged across laptops, CI pipelines, containerised Slurm clusters, and Tier-0 supercomputers.
-The project packages dependencies with Pixi [@pixi], submits workloads through Slurm using Dagster Pipes [@dagsterpipes], and streams logs plus scheduler metrics back to the Dagster UI.
+The project packages dependencies with Pixi @pixi, submits workloads through Slurm using Dagster Pipes @dagsterpipes, and streams logs plus scheduler metrics back to the Dagster UI.
 
 The key contribution is a unified compute resource (`ComputeResource`) that hides SSH transport (including password-only jump hosts and OTP prompts), dependency packaging, and queue configuration while still respecting Slurm’s scheduling semantics.
 Today the project ships with two production-ready execution modes—`local` for laptop/CI development and `slurm` for one-job-per-asset submissions.
@@ -51,11 +51,12 @@ Experimental support for session-based reuse and heterogeneous jobs is work in p
 
 # Statement of Need
 
-Research software engineers (RSE) and data scientists increasingly face cross-environment workflows on it's 22 edition the Research Software Engineering (RSE) International Survey [@hettrick_2022_7015772] every country marked HPC as important skill for RSE. 
-Reproducibility challenges in HPC environments are well-documented, [@ANTUNES2024100655] provided comprehensive coverage of issues unique to HPC in their survey.
-[@Courtes9882991] further examine the tension between reproducibility and performance, demonstrating that these goals need not be mutually exclusive.
+Research software engineers (RSE) and data scientists increasingly face cross-environment workflows.
+The 22nd edition of the Research Software Engineering (RSE) International Survey reports that every participating country now marks HPC as an important RSE skill @hettrick_2022_7015772.
+Reproducibility challenges in HPC environments are well documented, and Antunes et al. @ANTUNES2024100655 provide comprehensive coverage of issues unique to HPC in their survey.
+Courtes et al. @Courtes9882991 further examine the tension between reproducibility and performance, demonstrating that these goals need not be mutually exclusive.
 
-The 2024 Community Workshop on Practical Reproducibility in HPC produced a comprehensive report highlighting cost-effective reproducibility challenges and the need for tools that bridge development and production environments [@keahey_2025_15306610].
+The 2024 Community Workshop on Practical Reproducibility in HPC produced a comprehensive report highlighting cost-effective reproducibility challenges and the need for tools that bridge development and production environments @keahey_2025_15306610.
 Existing solutions either ignore the orchestrator (hand-written Slurm scripts) or bypass the scheduler entirely (running ad-hoc services).
 This leads to duplicated logic, fragile deployments, and a loss of telemetry. **dagster-slurm** was created to:
 
@@ -63,7 +64,7 @@ This leads to duplicated logic, fragile deployments, and a loss of telemetry. **
[…]
