+++ .github/workflows/library.yaml
@@ -64,7 +64,7 @@ jobs:
       - name: Install uv
         uses: astral-sh/setup-uv@v6
 
-      - uses: prefix-dev/setup-pixi@v0.9.2
+      - uses: prefix-dev/setup-pixi@v0.9.3
         with:
             pixi-version: v0.59.0
             cache: true
@@ -238,7 +238,7 @@ jobs:
       - name: Install uv
         uses: astral-sh/setup-uv@v6
 
-      - uses: prefix-dev/setup-pixi@v0.9.2
+      - uses: prefix-dev/setup-pixi@v0.9.3
         with:
             pixi-version: v0.59.0
             cache: true
@@ -343,7 +343,7 @@ jobs:
       - name: Install uv
         uses: astral-sh/setup-uv@v6
 
-      - uses: prefix-dev/setup-pixi@v0.9.2
+      - uses: prefix-dev/setup-pixi@v0.9.3
         with:
             pixi-version: v0.59.0
             cache: true
