+++ .github/workflows/tutorial.yaml
@@ -42,7 +42,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v5
 
-      - uses: prefix-dev/setup-pixi@v0.9.2
+      - uses: prefix-dev/setup-pixi@v0.9.3
         with:
             pixi-version: v0.59.0
             cache: true
