### 2025-11-21 5e4ad71
feat: tu datalab (#83)
* feat: upgrade deps

* feat: enable TU Wien datalab
