### 2025-11-22 48195c9
feat: refine slides
