### 2025-12-12 ea3fa18
feat: direct submission
