### 2025-12-12 983d6ad
feat: refine slides v2 (#84)
* feat: refine slides

* feat: deps upgrades

* feat: up upgrade

* feat: use prek

* feat: agent definitions
