+++ .github/workflows/tutorial.yaml
@@ -44,7 +44,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.9.3
         with:
-            pixi-version: v0.59.0
+            pixi-version: v0.61.0
             cache: true
             frozen: true
       - run: cd examples && pixi run -e dev --frozen lint
