### 2025-12-12 fa9023f
Revert "fix: try fixing tests?"
This reverts commit 8ffde97ce2df1b2d7e1eb2690fa9190a2868936e.
