### 2025-12-12 773dc6b
fix: refine for multi-asset
