+++ projects/dagster-slurm/dagster_slurm/resources/compute.py
@@ -1,5 +1,6 @@
 """Unified compute resource - main facade."""
 
+from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple
 
 from dagster import ConfigurableResource, InitResourceContext, get_dagster_logger
@@ -262,67 +263,88 @@ class ComputeResource(ConfigurableResource):
 
         return override
 
+    def _extract_asset_keys(self, context) -> list[Any]:
+        """Best-effort extraction of asset keys, including multi-asset outputs."""
+        asset_keys: list[Any] = []
+
+        # Primary asset key if available
+        try:
+            asset_key = getattr(context, "asset_key", None)
+            if asset_key:
+                asset_keys.append(asset_key)
+        except Exception:
+            pass
+
+        # Output names for multi-asset scenarios
+        output_names: list[str] = []
+        try:
+            output_names = list(
+                getattr(context, "selected_output_names", [])
+                or getattr(context, "output_names", [])
+            )
+        except Exception:
+            output_names = []
+
+        if not output_names:
+            try:
+                op_def = getattr(context, "op_def", None)
+                if op_def and getattr(op_def, "output_defs", None):
+                    output_names = [od.name for od in op_def.output_defs]
+            except Exception:
+                output_names = []
+
+        for output_name in output_names:
+            try:
+                key = context.asset_key_for_output(output_name)
+                if key:
+                    asset_keys.append(key)
+            except Exception:
+                continue
+
+        # Deduplicate while preserving order
+        seen = set()
+        unique_keys = []
+        for key in asset_keys:
+            try:
+                if key in seen:
+                    continue
+                seen.add(key)
+            except Exception:
+                # Non-hashable keys: keep them
+                pass
+            unique_keys.append(key)
+
+        return unique_keys
+
     def _should_force_env_push(self, context, explicit: Optional[bool]) -> bool:
         """Determine whether to force uploading the environment for this execution."""
         if explicit is not None:
             return explicit
 
-        try:
-            has_assets_def = getattr(context, "has_assets_def", False)
-            if callable(has_assets_def):
-                has_assets_def = has_assets_def()
+        # Check if context has asset definition
+        if not hasattr(context, "has_assets_def"):
+            return False
 
-            if has_assets_def and getattr(context, "assets_def", None):
-                metadata_by_key = getattr(context.assets_def, "metadata_by_key", {})  # type: ignore[attr-defined]
-                if not metadata_by_key:
-                    return False
+        has_assets_def = context.has_assets_def
+        if callable(has_assets_def):
+            has_assets_def = has_assets_def()
 
-                def _safe_asset_key() -> Optional[Any]:
-                    try:
-                        return getattr(context, "asset_key", None)
-                    except Exception:
-                        return None
+        if not (
+            has_assets_def and hasattr(context, "assets_def") and context.assets_def
+        ):
+            return False
 
-                asset_keys = []
-                asset_key = _safe_asset_key()
-                if asset_key:
-                    asset_keys.append(asset_key)
+        metadata_by_key = getattr(context.assets_def, "metadata_by_key", {})  # type: ignore[attr-defined]
+        if not metadata_by_key:
+            return False
 
-                output_names = []
-                try:
-                    output_names = list(
-                        getattr(context, "selected_output_names", [])
-                        or getattr(context, "output_names", [])
-                    )
-                except Exception:
-                    output_names = []
+        asset_keys = self._extract_asset_keys(context)
 
-                if not output_names:
-                    try:
-                        op_def = getattr(context, "op_def", None)
-                        if op_def and getattr(op_def, "output_defs", None):
-                            output_names = [od.name for od in op_def.output_defs]
-                    except Exception:
[…]
