### 2025-12-12 9464879
fix: reduce hasattr
