### 2025-12-13 afc6934
fix: a) configurable cache b) selective cache invalidation
