+++ projects/dagster-slurm/dagster_slurm/resources/compute.py
@@ -8,6 +8,7 @@ from pydantic import Field, PrivateAttr, model_validator
 from dagster._core.pipes.client import PipesClientCompletedInvocation
 
 from ..config.environment import ExecutionMode
+from ..config.runtime import SlurmRunConfig
 from ..helpers.ssh_pool import SSHConnectionPool
 from ..launchers.base import ComputeLauncher
 from ..managers.hetjob import HeterogeneousJobManager, HetJobComponent
@@ -99,6 +100,16 @@ class ComputeResource(ConfigurableResource):
     pre_deployed_env_path: Optional[str] = Field(
         default=None,
         description="If set, uses a pre-deployed environment at this path instead of live packaging.",
+    )
+
+    cache_inject_globs: Optional[List[str]] = Field(
+        default=None,
+        description=(
+            "List of --inject glob patterns whose file contents should affect the "
+            "environment cache key. If None, all --inject patterns from the pack command "
+            "are hashed. Use this to exclude workload-specific packages from cache "
+            "invalidation. Example: ['../dist/dagster_slurm-*.whl', 'projects/shared/dist/*.conda']"
+        ),
     )
 
     # Cluster reuse settings (session mode only)
@@ -211,6 +222,7 @@ class ComputeResource(ConfigurableResource):
                 auto_detect_platform=self.auto_detect_platform,
                 pack_platform=self.pack_platform,
                 pre_deployed_env_path=self.pre_deployed_env_path,
+                cache_inject_globs=self.cache_inject_globs,
             )
 
         elif self.mode == ExecutionMode.SLURM_SESSION:
@@ -228,6 +240,7 @@ class ComputeResource(ConfigurableResource):
                 auto_detect_platform=self.auto_detect_platform,
                 pack_platform=self.pack_platform,
                 pre_deployed_env_path=self.pre_deployed_env_path,
+                cache_inject_globs=self.cache_inject_globs,
             )
 
         else:  # ExecutionMode.SLURM_HETJOB
@@ -241,6 +254,7 @@ class ComputeResource(ConfigurableResource):
                 auto_detect_platform=self.auto_detect_platform,
                 pack_platform=self.pack_platform,
                 pre_deployed_env_path=self.pre_deployed_env_path,
+                cache_inject_globs=self.cache_inject_globs,
             )
 
     def _resolve_launcher(self, override: Optional[ComputeLauncher]) -> ComputeLauncher:
@@ -449,6 +463,7 @@ class ComputeResource(ConfigurableResource):
         force_env_push: Optional[bool] = None,
         skip_payload_upload: Optional[bool] = None,
         remote_payload_path: Optional[str] = None,
+        config: Optional[SlurmRunConfig] = None,
         **kwargs,
     ) -> PipesClientCompletedInvocation:
         """Execute asset with optional resource overrides.
@@ -469,12 +484,14 @@ class ComputeResource(ConfigurableResource):
                 - memory_gb: int
                 - framework: str ("ray" or "spark")
             force_env_push: Force repacking and uploading the environment for this run.
-                If not provided, falls back to asset metadata key 'force_slurm_env_push'.
+                If not provided, falls back to config.force_env_push or asset metadata.
             skip_payload_upload: If True, do not upload the payload script; expects the
-                remote payload to already exist. Can be set via asset metadata key
-                'skip_slurm_payload_upload'.
+                remote payload to already exist. Falls back to config.skip_payload_upload
+                or asset metadata key 'skip_slurm_payload_upload'.
             remote_payload_path: Remote path to an existing payload when skipping upload.
-                Can be set via asset metadata key 'slurm_payload_path'.
+                Falls back to config.remote_payload_path or asset metadata.
[…]
