+++ projects/dagster-slurm/dagster_slurm/helpers/env_packaging.py
@@ -1,5 +1,6 @@
 """Environment packaging using pixi pack."""
 
+import glob
 import re
 import shlex
 import subprocess
@@ -140,13 +141,76 @@ def pack_environment_with_pixi(
     return pack_file
 
 
+def _extract_inject_patterns(pack_cmd: Sequence[str]) -> list[str]:
+    """Extract --inject argument values from a pack command."""
+    patterns = []
+    i = 0
+    while i < len(pack_cmd):
+        arg = pack_cmd[i]
+        if arg == "--inject" and i + 1 < len(pack_cmd):
+            patterns.append(pack_cmd[i + 1])
+            i += 2
+        elif arg.startswith("--inject="):
+            patterns.append(arg[len("--inject=") :])
+            i += 1
+        else:
+            i += 1
+    return patterns
+
+
+def _resolve_and_hash_inject_files(
+    patterns: Sequence[str],
+    digest: "hashlib._Hash",
+    logger,
+) -> int:
+    """Resolve glob patterns and hash file contents. Returns count of files hashed."""
+    files_hashed = 0
+    for pattern in patterns:
+        matched_files = sorted(glob.glob(pattern))
+        if not matched_files:
+            logger.debug(f"No files matched inject pattern: {pattern}")
+            # Still include the pattern itself so different patterns yield different keys
+            digest.update(f"pattern:{pattern}".encode())
+            continue
+
+        for file_path in matched_files:
+            path = Path(file_path)
+            if path.is_file():
+                try:
+                    digest.update(path.read_bytes())
+                    files_hashed += 1
+                    logger.debug(f"Hashed inject file: {file_path}")
+                except OSError as e:
+                    logger.warning(f"Could not read inject file {file_path}: {e}")
+                    # Include path so missing files don't silently match
+                    digest.update(f"unreadable:{file_path}".encode())
+    return files_hashed
+
+
 def compute_env_cache_key(
-    pack_cmd: Sequence[str], lockfile: Path = Path("pixi.lock")
+    pack_cmd: Sequence[str],
+    lockfile: Path = Path("pixi.lock"),
+    cache_inject_globs: Optional[Sequence[str]] = None,
 ) -> Optional[str]:
     """Compute a stable cache key for the packed environment.
 
-    Uses the pixi lockfile contents and the pack command (which captures target platform)
-    to derive a short, reproducible identifier. Returns None if the lockfile is missing.
+    Uses the pixi lockfile contents, the pack command, and the contents of
+    injected files to derive a short, reproducible identifier.
+
+    Args:
+        pack_cmd: The pixi-pack command (e.g., ["pixi-pack", "--inject", "dist/*.whl", ...])
+        lockfile: Path to the pixi.lock file (default: pixi.lock in current directory)
+        cache_inject_globs: Optional list of inject patterns whose file contents should
+            affect the cache key. If None, all --inject patterns from pack_cmd are hashed.
+            Use this to exclude workload-specific packages that change frequently but
+            shouldn't invalidate the base environment cache.
+
+            Example: To only cache-invalidate on base library changes:
+                cache_inject_globs=["../dist/dagster_slurm-*.whl",
+                                    "projects/shared/dist/*.conda"]
+
+    Returns:
+        A 16-character hex string cache key, or None if lockfile is missing.
     """
     logger = get_dagster_logger()
 
@@ -158,8 +222,29 @@ def compute_env_cache_key(
         return None
 
     digest = hashlib.sha256()
+
+    # 1. Hash lockfile contents
     digest.update(lockfile.read_bytes())
+
+    # 2. Hash the pack command (captures platform, environment name, etc.)
     digest.update("::".join(pack_cmd).encode())
+
[…]
