+++ docs/docs/api/api_core.md
@@ -127,6 +127,8 @@ Heterogeneous job mode (optimal resource allocation):
 
 #### auto_detect_platform *: `bool`*
 
+#### cache_inject_globs *: `Optional`[`List`[`str`]]*
+
 #### cleanup_on_failure *: `bool`*
 
 #### cluster_reuse_tolerance *: `float`*
@@ -134,6 +136,8 @@ Heterogeneous job mode (optimal resource allocation):
 #### debug_mode *: `bool`*
 
 #### default_launcher *: `Annotated`[`Union`[[`ComputeLauncher`](#dagster_slurm.ComputeLauncher), `PartialResource`]]*
+
+#### default_skip_payload_upload *: `bool`*
 
 #### enable_cluster_reuse *: `bool`*
 
@@ -176,7 +180,7 @@ Register a newly created cluster for future reuse.
   * **gpus** (`int`) – Total GPUs in cluster
   * **memory_gb** (`int`) – Total memory in GB
 
-#### run(context, payload_path, launcher=None, extra_slurm_opts=None, resource_requirements=None, \*\*kwargs)
+#### run(context, payload_path, launcher=None, extra_slurm_opts=None, resource_requirements=None, force_env_push=None, skip_payload_upload=None, remote_payload_path=None, config=None, \*\*kwargs)
 
 Execute asset with optional resource overrides.
 
@@ -195,6 +199,15 @@ Execute asset with optional resource overrides.
     - gpus: int
     - memory_gb: int
     - framework: str (“ray” or “spark”)
+  * **force_env_push** (`Optional`[`bool`]) – Force repacking and uploading the environment for this run.
+    If not provided, falls back to config.force_env_push or asset metadata.
+  * **skip_payload_upload** (`Optional`[`bool`]) – If True, do not upload the payload script; expects the
+    remote payload to already exist. Falls back to config.skip_payload_upload
+    or asset metadata key ‘skip_slurm_payload_upload’.
+  * **remote_payload_path** (`Optional`[`str`]) – Remote path to an existing payload when skipping upload.
+    Falls back to config.remote_payload_path or asset metadata.
+  * **config** (`Optional`[[`SlurmRunConfig`](#dagster_slurm.SlurmRunConfig)]) – Optional SlurmRunConfig for run-time configuration via launchpad.
+    Values from config are used as defaults, but explicit parameters take precedence.
   * **\*\*kwargs** – Passed to client.run()
 * **Yields:**
   Dagster events
@@ -206,6 +219,21 @@ Execute asset with optional resource overrides.
 ```python
 # Simple execution with default resources
 yield from compute.run(context, "script.py")
+```
+
+```python
+# Using SlurmRunConfig for launchpad-configurable options
+@dg.asset
+def my_asset(
+    context: dg.AssetExecutionContext,
+    compute: ComputeResource,
+    config: SlurmRunConfig,
+):
+    return compute.run(
+        context=context,
+        payload_path="script.py",
+        config=config,
+    ).get_results()
 ```
 
 ```python
@@ -272,9 +300,9 @@ compute.run_hetjob(
 )
 ```
 
-#### session *: `Optional`[[`SlurmSessionResource`](#id43)]*
+#### session *: `Optional`[[`SlurmSessionResource`](#id45)]*
 
-#### slurm *: `Optional`[[`SlurmResource`](#id19)]*
+#### slurm *: `Optional`[[`SlurmResource`](#id21)]*
 
 #### teardown(context)
 
@@ -468,7 +496,7 @@ For proxy jumps, use the `_JUMP` suffix for jump host variables (e.g.,
[…]
