### 2025-12-17 f23845f
feat: use ty instead of pyright
