+++ examples/pyproject.toml
@@ -50,7 +50,7 @@ pytest = ">=8.3.4,<9"
 pytest-mock = ">=3.14.0,<4"
 pytest-cov = "~=6.0.0"
 ruff = ">=0.12.8,<1"
-pyrefly = ">=0.27.2,<1"
+ty = ">=0.0.2,<1"
 [tool.pixi.feature.ci-basics.pypi-dependencies]
 
 [tool.pixi.feature.dagster-basics.pypi-dependencies]
@@ -228,7 +228,7 @@ description = "Format python files"
 env = { RUST_LOG = "warn" }
 
 [tool.pixi.tasks.lint]
-cmd = "ruff check ./projects && yamllint -c yamllintconfig.yaml ./projects && taplo check && pyrefly check projects --project-excludes 'projects/*/.pixi' --python-interpreter-path $PIXI_PROJECT_ROOT/.pixi/envs/dev/bin/python3.12"
+cmd = "ruff check ./projects && yamllint -c yamllintconfig.yaml ./projects && taplo check && ty check ./projects"
 description = "Validate formatting and type check python files"
 
 [tool.pixi.tasks.test]
@@ -277,22 +277,17 @@ filterwarnings = [
 markers = ["data: tests using real input data"]
 
 
-[tool.pyrefly]
-project-includes = [
+[tool.ty]
+include = [
   "projects/dagster-slurm-example/dagster_slurm_example",
   "projects/dagster-slurm-example-shared/dagster_slurm_example_shared",
   "projects/dagster-slurm-example/dagster_slurm_example_tests",
 ]
-project-excludes = [
-  "**/build/**",
-  "**/.ruff_cache/**",
-  "**/.pixi/**",
-  "**/dist/**",
-]
+exclude = ["**/build/**", "**/.ruff_cache/**", "**/.pixi/**", "**/dist/**"]
 
-search-path = [
+src = [
   "projects/dagster-slurm-example",
   "projects/dagster-slurm-example-shared",
 ]
 
-python-version = "3.12.11"
+python-version = "3.12"
