### 2025-12-17 7bb7c49
fix: multi asset subselection
