### 2025-12-17 9512779
fix: caching and cache invalidation
