### 2025-12-19 97204ad
feat: direct submission (#85)
* feat: refine slides

* feat: deps upgrades

* feat: up upgrade

* feat: use prek

* feat: agent definitions

* feat: direct submission

* fix: new testing

* fix: try fixing tests?

* Revert "fix: try fixing tests?"

This reverts commit 8ffde97ce2df1b2d7e1eb2690fa9190a2868936e.

* fix: refine for multi-asset

* fix: ray lint; feat: allow asset specific ENV overrides

* fix: reduce hasattr

* fix: test

* fix: a) configurable cache b) selective cache invalidation

* fix: refine

* feat: use ty instead of pyright

* fix: multi asset subselection

* fix: caching an
[…]
