+++ projects/dagster-slurm/dagster_slurm/pipes_clients/slurm_pipes_client.py
@@ -21,7 +21,7 @@ from dagster import (
 )
 from dagster._core.pipes.client import PipesClientCompletedInvocation
 
-from ..helpers.env_packaging import pack_environment_with_pixi
+from ..helpers.env_packaging import compute_env_cache_key, pack_environment_with_pixi
 from ..helpers.message_readers import SSHMessageReader
 from ..helpers.metrics import SlurmMetricsCollector
 from ..helpers.ssh_helpers import TERMINAL_STATES
@@ -56,6 +56,7 @@ class SlurmPipesClient(PipesClient):
         auto_detect_platform: bool = True,
         pack_platform: Optional[str] = None,
         pre_deployed_env_path: Optional[str] = None,
+        cache_inject_globs: Optional[list[str]] = None,
     ):
         """Args:
         slurm_resource: Slurm cluster configuration
@@ -65,6 +66,10 @@ class SlurmPipesClient(PipesClient):
         debug_mode: If True, never cleanup files (for debugging)
         auto_detect_platform: Auto-detect platform (ARM vs x86) for pixi pack
         pack_platform: Override platform ('linux-64', 'linux-aarch64', 'osx-arm64').
+        cache_inject_globs: Optional list of --inject glob patterns whose file contents
+            should affect the environment cache key. If None, all --inject patterns
+            from the pack command are hashed. Use this to exclude workload-specific
+            packages from cache invalidation (e.g., only include base libraries).
 
         """
         super().__init__()
@@ -82,6 +87,7 @@ class SlurmPipesClient(PipesClient):
         self._ssh_pool = None
         self._cancellation_requested = False
         self.pre_deployed_env_path = pre_deployed_env_path
+        self.cache_inject_globs = cache_inject_globs
 
     def run(  # type: ignore[override]
         self,
@@ -92,6 +98,11 @@ class SlurmPipesClient(PipesClient):
         extras: Optional[Dict[str, Any]] = None,
         use_session: bool = False,
         extra_slurm_opts: Optional[Dict[str, Any]] = None,
+        force_env_push: Optional[bool] = None,
+        skip_payload_upload: Optional[bool] = None,
+        remote_payload_path: Optional[str] = None,
+        pack_cmd_override: Optional[list[str]] = None,
+        pre_deployed_env_path_override: Optional[str] = None,
         **kwargs,
     ) -> PipesClientCompletedInvocation:
         """Execute payload on Slurm cluster with real-time log streaming.
@@ -104,6 +115,12 @@ class SlurmPipesClient(PipesClient):
             extras: Extra data to pass via Pipes
             use_session: If True and session_resource provided, use shared allocation
             extra_slurm_opts: Override Slurm options (non-session mode)
+            force_env_push: If True, always repack and upload the environment even when
+                a cached copy exists for the current lockfile/pack command.
+            skip_payload_upload: If True, do not upload the payload script (assumes
+                it already exists remotely).
+            remote_payload_path: Optional pre-existing remote payload path to use when
+                skipping upload.
             **kwargs: Additional arguments (ignored, for forward compatibility)
 
         Yields:
@@ -117,6 +134,10 @@ class SlurmPipesClient(PipesClient):
                 "Context is not part of a Dagster run, generating a temporary run_id."
             )
             run_id = uuid.uuid4().hex
+        force_env_push = bool(force_env_push) if force_env_push is not None else False
+        skip_payload_upload = (
+            bool(skip_payload_upload) if skip_payload_upload is not None else False
+        )
 
         # Setup SSH connection pool
         ssh_pool = SSHConnectionPool(self.slurm.ssh)
@@ -140,63 +161,42 @@ class SlurmPipesClient(PipesClient):
                 # Store control path for log streaming
[…]
