### 2025-12-19 bf76a4e
1.8.0-rc.1 [skip ci]
Automatically generated by python-semantic-release
