### 2025-12-19 e9599a9
fix: direct lint env path
