### 2025-12-19 0b0cb3e
fix: uv without pixi
