### 2025-12-19 45673ce
fix: add ty directly from pixi
