+++ .github/workflows/library.yaml
@@ -62,7 +62,7 @@ jobs:
         uses: actions/checkout@v5
 
       - name: Install uv
-        uses: astral-sh/setup-uv@v6
+        uses: astral-sh/setup-uv@v7
 
       - uses: prefix-dev/setup-pixi@v0.9.3
         with:
@@ -236,7 +236,7 @@ jobs:
           token: ${{ steps.generate_token.outputs.token }}
 
       - name: Install uv
-        uses: astral-sh/setup-uv@v6
+        uses: astral-sh/setup-uv@v7
 
       - uses: prefix-dev/setup-pixi@v0.9.3
         with:
@@ -341,7 +341,7 @@ jobs:
           fi
 
       - name: Install uv
-        uses: astral-sh/setup-uv@v6
+        uses: astral-sh/setup-uv@v7
 
       - uses: prefix-dev/setup-pixi@v0.9.3
         with:
