### 2025-12-19 b3f803a
fix: try pre-creating env
