+++ .github/workflows/library.yaml
@@ -69,7 +69,7 @@ jobs:
             pixi-version: v0.61.0
             cache: true
             frozen: true
-      - run: pixi run -e build --frozen lint
+      - run: pixi install -e build && pixi run -e build --frozen lint
 
       - name: Run unit tests without real SLURM
         run: pixi run -e build --frozen testpixi
