### 2025-12-19 4a20ec2
1.9.0 [skip ci]
Automatically generated by python-semantic-release
