### 2026-01-12 f14d1ff
fix: musica, allow ulimit and other ray startup customization
