+++ projects/dagster-slurm/dagster_slurm/launchers/ray.py
@@ -36,6 +36,10 @@ class RayLauncher(ComputeLauncher):
     object_store_memory_gb: Optional[int] = Field(
         default=None, description="Object store size (None = auto)"
     )
+    ray_start_args: list[str] = Field(
+        default_factory=list,
+        description="Extra arguments to pass to ray start on the head node.",
+    )
     redis_password: Optional[str] = Field(
         default=None, description="Redis password (None = auto-generate with uuidgen)"
     )
@@ -48,6 +52,10 @@ class RayLauncher(ComputeLauncher):
     )
     worker_startup_delay: int = Field(
         default=1, description="Seconds between worker starts"
+    )
+    pre_start_commands: list[str] = Field(
+        default_factory=list,
+        description="Optional shell commands to run before ray start (e.g., ulimit).",
     )
     worker_cpu_bind: Optional[str] = Field(
         default=None,
@@ -67,6 +75,17 @@ class RayLauncher(ComputeLauncher):
         default="hash_jobid",
         description="'fixed' or 'hash_jobid' for head/dashboard ports.",
     )
+
+    def _render_override_block(self, date_fmt: str) -> str:
+        if not self.pre_start_commands:
+            return ""
+        lines = []
+        for command in self.pre_start_commands:
+            lines.append(
+                f'echo "[$({date_fmt})] Applying pre-start command: {command}"'
+            )
+            lines.append(command)
+        return "\n    ".join(lines)
 
     def prepare_execution(
         self,
@@ -152,7 +171,9 @@ class RayLauncher(ComputeLauncher):
     else
         echo "[$({date_fmt})] Single-node mode detected. Starting local Ray cluster..."
     """
-            local_lines = self._generate_local_template(date_fmt, activation_script)
+            local_lines = self._generate_local_template(
+                date_fmt, activation_script, working_dir
+            )
             script += local_lines
 
             script += f"""
@@ -175,9 +196,12 @@ class RayLauncher(ComputeLauncher):
         )
 
     def _generate_local_template(
-        self, date_fmt: str, activation_script: Optional[str]
+        self, date_fmt: str, activation_script: Optional[str], working_dir: str
     ) -> str:
         """Generate Ray startup for local (single-node) mode."""
+        override_block = self._render_override_block(date_fmt)
+        if override_block:
+            override_block = f"    {override_block}\n"
         # Build object store argument if specified
         obj_store = ""
         if self.object_store_memory_gb is not None:
@@ -202,7 +226,34 @@ class RayLauncher(ComputeLauncher):
         port=$(( {self.ray_port} + off ))
         dash_port=$(( {self.dashboard_port} + off ))
     fi
-    # Start local Ray cluster
+    # Resolve head address for local mode (prefer Slurm node IP when available)
+    head_bind_addr="127.0.0.1"
+    if [[ -n "${{SLURM_JOB_ID:-}}" && "{str(self.use_head_ip).lower()}" == "true" ]]; then
+      head_node_name="$(hostname)"
+      if command -v getent >/dev/null 2>&1; then
+        ipv4=$(getent ahostsv4 "$head_node_name" | awk 'NR==1{{print $1}}')
+        if [[ -n "$ipv4" ]]; then
+          head_bind_addr="$ipv4"
+        else
+          ipv6=$(getent ahostsv6 "$head_node_name" | awk 'NR==1{{print $1}}')
+          if [[ -n "$ipv6" ]]; then head_bind_addr="$ipv6"; fi
[…]
