### 2026-01-12 40b6065
feat: musica (#88)
* enable support for musica by configurable pre-start commands

---------

Co-authored-by: Martin Pfister <martin.pfister@tuwien.ac.at>
Co-authored-by: Georg Heiler <georg.kf.heiler@gmail.com>
