### 2026-01-12 6b969d4
1.9.0-rc.1 [skip ci]
Automatically generated by python-semantic-release
