### 2026-01-12 20c108a
1.10.0 [skip ci]
Automatically generated by python-semantic-release
