+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <!-- version list -->
 
+## v1.10.0 (2026-01-12)
+
+
 ## v1.9.0-rc.1 (2026-01-12)
 
 ### Features
