### 2026-01-19 84a10b3
fix:remove https prefix for dois (#89)
