+++ docs/paper.bib
@@ -79,7 +79,7 @@
       archivePrefix={arXiv},
       primaryClass={cs.CL},
       url={https://arxiv.org/abs/2408.09869},
-      doi={https://doi.org/10.48550/arXiv.2408.09869} 
+      doi={10.48550/arXiv.2408.09869} 
 }
 
 @software{dagster,
@@ -100,7 +100,7 @@
   pages = {44--60},
   year = {2003},
   publisher = {Springer},
-  doi = {https://doi.org/10.1007/10968987_3}
+  doi = {10.1007/10968987_3}
 }
 
 @software{pixi,
@@ -126,7 +126,7 @@ volume = {53},
 pages = {100655},
 year = {2024},
 issn = {1574-0137},
-doi = {https://doi.org/10.1016/j.cosrev.2024.100655},
+doi = {10.1016/j.cosrev.2024.100655},
 url = {https://www.sciencedirect.com/science/article/pii/S157401372400039X},
 author = {Benjamin Antunes and David R.C. Hill},
 keywords = {Reproducibility, Reproducible research, Computational science, Replicability, Repeatability, High performance computing},
@@ -141,7 +141,7 @@ abstract = {Reproducibility is widely acknowledged as a fundamental principle in
   number={3},
   pages={77-80},
   keywords={High performance computing;Research and development;Reproducibility of results},
-  doi={https://doi.org/10.1109/MCSE.2022.3165626}}
+  doi={10.1109/MCSE.2022.3165626}}
 @misc{hayotsasson2025address,
       title={Addressing Reproducibility Challenges in HPC with Continuous Integration}, 
       author={Valérie Hayot-Sasson and Nathaniel Hudson and André Bauer and Maxime Gonthier and Ian Foster and Kyle Chard},
@@ -150,7 +150,7 @@ abstract = {Reproducibility is widely acknowledged as a fundamental principle in
       archivePrefix={arXiv},
       primaryClass={cs.DC},
       url={https://arxiv.org/abs/2508.21289},
-      doi={https://doi.org/10.48550/arXiv.2508.21289},
+      doi={10.48550/arXiv.2508.21289},
 }
 @misc{keahey_2025_15306610,
   author       = {Keahey, Kate and
@@ -177,5 +177,5 @@ abstract = {Reproducibility is widely acknowledged as a fundamental principle in
       archivePrefix={arXiv},
       primaryClass={cs.SE},
       url={https://arxiv.org/abs/2507.16166}, 
-      doi={https://doi.org/10.48550/arXiv.2507.16166},
+      doi={10.48550/arXiv.2507.16166},
 }
