+++ .github/workflows/library.yaml
@@ -66,7 +66,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.9.3
         with:
-            pixi-version: v0.61.0
+            pixi-version: v0.63.2
             cache: true
             frozen: true
       - run: pixi install -e build && pixi run -e build --frozen lint
@@ -240,7 +240,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.9.3
         with:
-            pixi-version: v0.61.0
+            pixi-version: v0.63.2
             cache: true
             frozen: true
 
@@ -345,7 +345,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.9.3
         with:
-            pixi-version: v0.61.0
+            pixi-version: v0.63.2
             cache: true
             frozen: true
 
