### 2026-01-20 2948646
feat: 1st steps towards docling demo
