+++ examples/projects/dagster-slurm-example/dagster_slurm_example/defs/ray/docling_assets.py
@@ -1,0 +1,202 @@
+"""Assets for distributed document processing using docling and Ray.
+
+This example demonstrates how to use dagster-slurm with Ray for large-scale
+document processing workflows. In development, it runs locally; in production
+on Slurm clusters, it spawns a multi-node Ray cluster for parallel processing.
+"""
+
+import dagster as dg
+from dagster_slurm import ComputeResource, RayLauncher, SlurmRunConfig
+
+
+@dg.asset(
+    description="Process PDF documents using docling with distributed Ray workers",
+)
+def process_documents_with_docling(
+    context: dg.AssetExecutionContext,
+    compute_ray: ComputeResource,
+    config: SlurmRunConfig,
+):
+    """Process PDF documents using docling and Ray for parallel conversion.
+
+    This asset demonstrates:
+    - Distributed document processing with Ray Data
+    - Converting PDFs to markdown using docling
+    - Seamless local dev vs Slurm production execution
+    - Multi-node Ray cluster orchestration on HPC
+
+    In local mode: Starts a local Ray instance
+    In Slurm mode: Spawns Ray cluster across allocated nodes
+
+    Args:
+        context: Dagster execution context
+        compute_ray: ComputeResource configured with RayLauncher
+        config: Slurm run configuration
+    """
+    script_path = dg.file_relative_path(
+        __file__,
+        "../../../../dagster-slurm-example-hpc-workload/dagster_slurm_example_hpc_workload/ray/process_documents_docling.py",
+    )
+
+    # Configure Ray launcher for document processing
+    # Adjust num_nodes and resources based on your workload
+    ray_launcher = RayLauncher(
+        num_gpus_per_node=0,  # Set to >0 if using GPU-accelerated OCR
+    )
+
+    context.log.info("Starting distributed document processing with docling...")
+
+    completed_run = compute_ray.run(
+        context=context,
+        payload_path=script_path,
+        config=config,
+        launcher=ray_launcher,
+        extra_env={
+            # Document processing configuration
+            "INPUT_GLOB": "data/**/*.pdf",  # Adjust to your data location
+            "OUTPUT_DIR": "/tmp/docling_output",  # Output directory for processed docs
+            "NUM_WORKERS": "2",  # Number of parallel workers
+            "BATCH_SIZE": "4",  # Documents per batch
+        },
+        extra_slurm_opts={
+            # Resource allocation for document processing
+            # Single node = local Ray mode
+            # Multi-node = distributed Ray cluster
+            "nodes": 1,  # Increase for larger workloads
+            "cpus_per_task": 4,  # CPUs per Ray worker
+            "mem": "8G",  # Memory per node
+            # Uncomment for GPU-accelerated OCR:
+            # "gres": "gpu:1",
+        },
+    )
+
+    yield from completed_run.get_results()
+
+
+@dg.asset(
+    deps=[process_documents_with_docling],
+    description="Aggregate and analyze docling processing results",
+)
+def analyze_docling_results(  # noqa: C901
+    context: dg.AssetExecutionContext,
+):
+    """Analyze results from document processing and generate summary statistics.
+
+    This asset demonstrates downstream processing of docling results,
+    showing how to chain multiple assets in a document processing pipeline.
+
+    Args:
+        context: Dagster execution context
+    """
+    # Get metadata from upstream asset
+    event_record = context.instance.get_latest_materialization_event(
+        asset_key=process_documents_with_docling.key
+    )
+
+    if (
+        not event_record
+        or not event_record.asset_materialization
+        or not event_record.asset_materialization.metadata
+    ):
+        raise dg.DagsterError(
+            f"Could not find metadata from upstream asset '{process_documents_with_docling.key}'. "
[…]
