+++ examples/projects/dagster-slurm-example-hpc-workload/dagster_slurm_example_hpc_workload/ray/process_documents_docling.py
@@ -1,0 +1,421 @@
+"""Docling document processing with Ray for distributed PDF conversion.
+
+This example processes PDF documents using docling and Ray Data for parallel processing.
+Adapted for dagster-slurm integration to demonstrate HPC document processing workflows.
+
+Based on: https://github.com/l-mds/duckpond/blob/main/projects/100_combined/ai_example/
+"""
+
+import time
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any, Dict, List, Optional
+
+import ray
+import ray.data as rd
+from dagster_pipes import PipesContext, open_dagster_pipes
+from docling.datamodel.base_models import InputFormat
+from docling.datamodel.document import ConversionResult, ConversionStatus
+from docling.document_converter import DocumentConverter
+from docling_core.types.doc import ImageRefMode
+
+# Optional: Advanced OCR with RapidOCR
+# Uncomment these imports to enable RapidOCRDocumentConverter
+# from docling.datamodel.accelerator_options import AcceleratorDevice, AcceleratorOptions
+# from docling.datamodel.pipeline_options import RapidOcrOptions, ThreadedPdfPipelineOptions
+
+
+@dataclass
+class ConvertSummary:
+    """Summary of document conversion results."""
+
+    ok: bool
+    src_path: str
+    pages: Optional[int]
+    elapsed_s: float
+    output_dir: Optional[str]
+    error: Optional[str] = None
+
+
+class BasicDocumentConverter:
+    """Basic docling converter for PDF documents.
+
+    This is the simple converter that works out-of-the-box with minimal configuration.
+    For more advanced OCR capabilities, use RapidOCRDocumentConverter below.
+
+    Based on: https://github.com/l-mds/duckpond/blob/main/projects/100_combined/ai_example/ai_example/defs/document_ai/document_plain_simple.py
+    """
+
+    def __init__(self, document_format: InputFormat = InputFormat.PDF):
+        self.converter = DocumentConverter()
+        self.converter.initialize_pipeline(document_format)
+
+    def convert_one(self, uri: str) -> ConversionResult:
+        """Convert a single document."""
+        return self.converter.convert(uri)
+
+
+# Uncomment this class to use advanced RapidOCR-based conversion
+# You'll also need to uncomment the imports at the top of this file
+#
+# class RapidOCRDocumentConverter:
+#     """Advanced docling converter with RapidOCR for better OCR quality.
+#
+#     This converter uses RapidOCR models for higher quality text extraction
+#     from scanned documents and images within PDFs.
+#
+#     Based on: https://github.com/l-mds/duckpond/blob/main/projects/100_combined/ai_example/ai_example/defs/document_ai/document_plain_rapidocr.py
+#
+#     Requirements:
+#     - RapidOCR model files (see rapidocr_models_root configuration)
+#     - Additional dependencies: psutil
+#     """
+#
+#     def __init__(self, models_root: Optional[str] = None):
+#         self.format = InputFormat.PDF
+#         self.models_root = models_root or os.getenv("RAPIDOCR_MODELS_ROOT", "/path/to/models")
+#         self.converter = DocumentConverter(
+#             format_options={
+#                 self.format: PdfFormatOption(
+#                     pipeline_options=self._configure_converter()
+#                 )
+#             }
+#         )
+#         self.converter.initialize_pipeline(self.format)
+#
+#     def _configure_converter(self):
+#         """Configure the converter with RapidOCR options."""
+#         pipeline_options = ThreadedPdfPipelineOptions()
+#         pipeline_options.do_ocr = True
+#         pipeline_options.images_scale = 2
+#         pipeline_options.generate_page_images = False
+#         pipeline_options.do_picture_classification = True
+#         pipeline_options.generate_picture_images = True
+#
[…]
