### 2026-01-20 3ff33c3
feat: ARM runner for slurm image
