+++ examples/pyproject.toml
@@ -10,6 +10,11 @@ requires-python = ">=3.12,<3.13"
 platforms = ["linux-64", "osx-arm64", "linux-aarch64"]
 channels = ["conda-forge"]
 
+[tool.pixi.pypi-options]
+# Prefer binary wheels over building from source
+# needed for docling apparently
+extra-index-urls = []
+
 [build-system]
 build-backend = "hatchling.build"
 requires = ["hatchling"]
