+++ projects/dagster-slurm/README.md
@@ -27,7 +27,7 @@ Example
 
 ```bash
 git clone https://github.com/ascii-supply-networks/dagster-slurm.git
-docker compose up -d --build
+docker compose up
 cd dagster-slurm/examples
 ```
 
