### 2026-01-21 fae1ce4
feat: make configurable
