### 2026-01-21 8c98bc9
fix: robustify deploy script
