### 2026-01-21 0d98695
fix: add more logging
