+++ examples/scripts/deploy_environment.py
@@ -3,6 +3,7 @@ import os
 import subprocess
 import sys
 from loguru import logger
+import shlex
 import json
 
 from dagster_slurm import SSHConnectionResource
@@ -42,12 +43,12 @@ def main():
         default="deployment_metadata.json",
     )
     args = parser.parse_args()  
-    env_path = ".env" 
+    env_path = ".env"
     if os.path.exists(env_path):
         dotenv.load_dotenv(env_path)
-        print(f"Environment loaded from {env_path}")
+        logger.info(f"Environment loaded from {env_path}")
     else:
-        print(f"No .env file found at {env_path}")
+        logger.info(f"No .env file found at {env_path} (skipping).")
     try:
         base_path = os.environ["SLURM_DEPLOYMENT_BASE_PATH"]
         ssh_host = os.environ["SLURM_EDGE_NODE_HOST"]
@@ -62,12 +63,17 @@ def main():
         if key_path and password:
             raise ValueError("Authentication conflict: You cannot set both SLURM_EDGE_NODE_KEY_PATH and SLURM_EDGE_NODE_PASSWORD.")
 
+        extra_opts = shlex.split(os.getenv("SLURM_EDGE_NODE_OPTS_EXTRA", ""))
+        if extra_opts:
+            logger.info(f"Using extra SSH options: {extra_opts}")
+
         ssh_config = SSHConnectionResource(
             host=ssh_host,
             port=ssh_port,
             user=ssh_user,
             key_path=key_path,
             password=password,
+            extra_opts=extra_opts,
         )
         auth_method = "key-based" if key_path else "password-based"
         logger.info(f"Loaded SSH configuration for {ssh_config.user}@{ssh_config.host} (using {auth_method} auth).")
