### 2026-01-21 6ace264
fix: check if small runner works
