### 2026-01-21 3d83585
fix: refine many aspects of env handling, docling still broken
