### 2026-01-21 12e7825
fix: dockling env handling
