### 2026-01-21 #9795 @junghans
> **Paper file info**:
>
> ⚠️ Wordcount for `paper.md` is **1210**

This is a bit longer than the recommended 250-1000 words, so we will have to shorten the draft a bit.
