### 2026-01-21 be95009
feat: explore fix
