### 2026-01-21 0161dae
feat: explore fix
