### 2026-01-21 b9f9ada
feat: explore fix
