+++ examples/scripts/pack_environment.py
@@ -95,11 +95,18 @@ def main() -> int:
         if not args.build_missing:
             raise
         print(f"{exc} -> building missing artifacts")
+        build_env = os.environ.copy()
+        build_env["PIXI_PROJECT_MANIFEST"] = str(base_dir / "pyproject.toml")
         for task in BUILD_TASKS:
             build_cmd = ["pixi", "run", "--frozen", task]
             print(f"Running build task: {' '.join(build_cmd)}")
             build_result = subprocess.run(
-                build_cmd, check=False, cwd=base_dir, capture_output=True, text=True
+                build_cmd,
+                check=False,
+                cwd=base_dir,
+                capture_output=True,
+                text=True,
+                env=build_env,
             )
             if build_result.stdout:
                 print(build_result.stdout)
