+++ .github/workflows/library.yaml
@@ -70,6 +70,13 @@ jobs:
             pixi-version: v0.63.2
             cache: true
             frozen: true
+      - name: Early pack smoke check
+        env:
+          PYTHONUNBUFFERED: "1"
+        run: |
+          set -euo pipefail
+          cd examples
+          pixi run -e opstooling --frozen python scripts/pack_environment.py --env packaged-cluster --platform auto --build-missing --debug
       - run: pixi install -e build && pixi run -e build --frozen lint
 
       - name: Run unit tests without real SLURM
@@ -201,6 +208,9 @@ jobs:
             echo "Listing examples/ and dist/ for context..."
             ls -la
             ls -la dist || true
+            ls -la projects || true
+            ls -la projects/*/dist || true
+            ls -la projects/*/dist/*.conda || true
             ls -la ../dist || true
             exit 1
           fi
