+++ examples/scripts/pack_environment.py
@@ -1,3 +1,29 @@
+#!/usr/bin/env python3
+"""
+Pack a pixi environment with local dependencies injected.
+
+This script builds local packages and injects them into a pixi-pack environment.
+It's designed to be placed in scripts/pack_environment.py in your pixi workspace.
+
+Usage:
+    python scripts/pack_environment.py --env packaged-cluster --platform linux-64
+    python scripts/pack_environment.py --build-missing --debug
+
+Directory structure assumed:
+    your-workspace/
+    ├── pyproject.toml          # Main pixi project
+    ├── scripts/
+    │   └── pack_environment.py # This script
+    └── projects/               # Local packages to build and inject
+        ├── your-package-1/
+        ├── your-package-2/
+        └── ...
+
+To adapt for your project:
+1. Update INJECT_PATTERNS below to match your local packages
+2. Update BUILD_COMMANDS to define how to build each package
+3. All paths are relative to the workspace root (where pyproject.toml lives)
+"""
 import argparse
 import os
 import platform
@@ -5,18 +31,47 @@ import subprocess
 from pathlib import Path
 import time
 
+# ============================================================================
+# CONFIGURATION - Customize for your project structure
+# ============================================================================
 
+# Patterns for built artifacts to inject into the packed environment
+# Paths are relative to the workspace root (where pyproject.toml lives)
+#
+# Example for a simple project with just one package at root:
+#   INJECT_PATTERNS = ["dist/my_package-*.conda"]
+#
+# Example for multiple packages in a projects/ directory:
+#   INJECT_PATTERNS = [
+#       "projects/package-a/dist/package_a-*.conda",
+#       "projects/package-b/dist/package_b-*.whl",
+#   ]
 INJECT_PATTERNS = [
+    # Local conda packages
     "projects/dagster-slurm-example-shared/dist/dagster_slurm_example_shared-*.conda",
-    "../dist/dagster_slurm-*-py3-none-any.whl",
     "projects/dagster-slurm-example-hpc-workload/dist/dagster_slurm_example_hpc_workload-*.conda",
     "projects/dagster-slurm-example/dist/dagster_slurm_example-*.conda",
+    # External dependency (dagster-slurm library from parent repo) - remove if not needed
+    "../dist/dagster_slurm-*-py3-none-any.whl",
 ]
 
+# Build commands: (relative_dir, command_list)
+# Commands are executed with cwd set to relative_dir (relative to workspace root)
+#
+# Example for a simple project at root:
+#   BUILD_COMMANDS = [(".", ["pixi", "build", "-o", "dist"])]
+#
+# Example for multiple packages:
+#   BUILD_COMMANDS = [
+#       ("projects/package-a", ["pixi", "build", "-o", "dist"]),
+#       ("projects/package-b", ["uv", "build", "-o", "dist"]),
+#   ]
 BUILD_COMMANDS = [
+    # Build local conda packages
     ("projects/dagster-slurm-example-shared", ["pixi", "build", "-o", "dist"]),
     ("projects/dagster-slurm-example-hpc-workload", ["pixi", "build", "-o", "dist"]),
     ("projects/dagster-slurm-example", ["pixi", "build", "-o", "dist"]),
+    # Build external dependency (dagster-slurm library from parent repo) - remove if not needed
     ("../projects", ["pixi", "run", "-e", "build", "--frozen", "build-lib"]),
 ]
 
@@ -32,6 +87,7 @@ def _detect_platform() -> str:
 
 
 def _resolve_inject_args(base_dir: Path, allow_missing: bool) -> list[str]:
+    """Resolve inject patterns to actual file paths."""
     args: list[str] = []
     for pattern in INJECT_PATTERNS:
         matches = sorted(
[…]
