+++ docs/docs/how-to/hpc-musica.md
@@ -3,84 +3,99 @@ sidebar_position: 99
 title: HPC - MUSICA
 ---
 
-Login on MUSICA is either via a link / qr-code on each login (not feasible for dagster-slurm) or with the help of the step client.
-To obtain a time-limited ssh certificate that can be used with dagster-slurm, use the step client. Execute the following command and log in in your web browser:
+MUSICA requires SSH certificate authentication via the Step client. Standard password or key-based authentication is not supported.
+
+## Authentication
+
+Generate a time-limited SSH certificate:
 
 ```bash
-step ssh certificate 'your.email.address@example.com' ~/.ssh/id_musica --no-password --insecure --force
+step ssh certificate 'your.email@example.com' ~/.ssh/id_musica --no-password --insecure --force
 ```
 
-## Optional: automated Step/OIDC refresh
+After running this command, authenticate in your browser. The certificate is valid for a limited time.
 
-The example repo supports an automated refresh flow that fetches an OIDC token and runs
-`step ssh login` before each SSH connection when `SLURM_SUPERCOMPUTER_SITE=musica`.
-Set the following variables (in addition to the SSH settings below):
+### Automatic Certificate Refresh (Recommended)
+
+Enable automatic OIDC-based certificate refresh by setting `SLURM_SUPERCOMPUTER_SITE=musica` and adding these variables:
 
 ```dotenv
-# OIDC client credentials (app password is valid for 30 days)
+# OIDC credentials (app password valid for 30 days)
 ASC_OIDC_CLIENT_ID=...
 ASC_OIDC_USERNAME=xy12345
 ASC_OIDC_APP_PASSWORD=...
 ASC_OIDC_TOKEN_URL=https://auth.asc.ac.at/application/o/token/
-ASC_OIDC_SCOPE=profile
 
-# Step settings
-ASC_STEP_CONTEXT=asc
+# Step CA configuration
 ASC_STEP_CA_URL=https://auth.asc.ac.at:9000
 ASC_STEP_FINGERPRINT=44b048473242281db1da57124c2b843741d6c92a8fb5d0482dec032e957f2919
-ASC_STEP_CERT_PATH=~/.step/ssh/certs/asc.crt
-ASC_STEP_REFRESH_SKEW=30
 ```
 
-You can verify the refresh flow locally with:
+The refresh hook automatically fetches new certificates before SSH connections.
+
+**Verify the flow:**
 
 ```bash
-cd examples
-pixi exec -e dev -- python scripts/refresh_step_auth.py
+pixi run -e dev python scripts/refresh_step_auth.py
 ```
 
-To rotate the app password via the authentik API (requires a separate API token with `goauthentik.io/api` scope):
+**Rotate app password** (requires API token with `goauthentik.io/api` scope):
 
 ```bash
-cd examples
-pixi exec -e dev -- python scripts/refresh_step_auth.py --rotate-app-password
+pixi run -e dev python scripts/refresh_step_auth.py --rotate-app-password
 ```
 
-Required API env vars for rotation:
+Required for rotation:
 
 ```dotenv
 ASC_AUTHENTIK_API_BASE=https://auth.asc.ac.at
-ASC_AUTHENTIK_API_TOKEN=...  # API token with goauthentik.io/api scope
+ASC_AUTHENTIK_API_TOKEN=...
 ASC_AUTHENTIK_TOKEN_IDENTIFIER=dagster-slurm
-ASC_AUTHENTIK_TOKEN_INTENT=app_password
-ASC_AUTHENTIK_TOKEN_EXPIRES_DAYS=30
-ASC_AUTHENTIK_APP_PASSWORD_OUTPUT=~/.secrets/asc_app_password.txt
 ```
 
-## Sample configuration: MUSICA
[…]
