### 2026-01-22 a2f6679
feat: refine docs + format
