+++ docs/docs/applications/document-preprocessing-docling.md
@@ -5,10 +5,10 @@ title: Document Preprocessing with Docling
 
 # Document Preprocessing with Docling
 
-Process PDF documents at scale using [docling](https://github.com/DS4SD/docling) and Ray. This example demonstrates distributed document processing on HPC clusters with automatic Ray cluster management.
+Process PDF documents at scale using [docling](https://github.com/DS4SD/docling) and [Ray](https://ray.io/). This example demonstrates distributed document processing on HPC clusters with automatic Ray cluster management.
 
-:::info Reference Implementation
-Adapted from [duckpond's docling+Ray example](https://github.com/l-mds/duckpond/tree/main/projects/100_combined/ai_example/ai_example/defs/document_ai) for HPC/Slurm environments.
+:::info Ray Integration
+For a complete guide on using Ray with dagster-slurm, including cluster management and distributed computing patterns, see [Ray on Slurm](../integration-ray/ray.md).
 :::
 
 ## Overview
@@ -109,14 +109,11 @@ Key components:
 
 ## Usage
 
-The payload script supports two execution modes:
+The application supports three execution modes with increasing complexity. **Start simple without Dagster, then add orchestration, then scale to HPC.**
 
 ### Test Data
 
-Sample PDFs are included in `examples/data/` for testing:
-
-- **WO2021041671-eval-small.pdf** - 4-page patent document (596KB)
-- **test-document-{1-5}.pdf** - 5 copies for batch processing tests
+Sample PDFs are included in `examples/data/` for testing all three modes.
 
 For HPC testing, transfer data to your cluster:
 
@@ -131,11 +128,59 @@ rsync -avz --progress examples/data/ user@hpc-cluster:/path/on/hpc/
 docker cp examples/data/. slurmctld:/home/submitter/dagster-slurm-data/
 ```
 
-### Mode 1: Via Dagster (Pipes Mode)
+:::warning AI Model Downloads
+Docling and other AI workloads require downloading models on first run. **Network connectivity requirements are cluster-dependent:**
 
-This is the primary integration mode for production workflows.
+- **Local development**: Models download automatically
+- **HPC clusters with internet access**: Models download to compute nodes (may be slow)
+- **Air-gapped/restricted clusters**: Pre-download required via edge node or proxy
 
-#### Local Development
+See [Pre-Downloading Models](#pre-downloading-models-for-restricted-clusters) below for setup instructions.
+:::
+
+### 1. Local Python Script
+
+**Simplest:** Run the processing script directly as a standalone Python script. No orchestration, no slurm —just Python, Ray, and docling.
+
+**Requirements:** Python environment with Ray and docling.
+
+```bash
+# Navigate to the payload script directory
+cd examples/
+pixi shell -e dev
+cd projects/dagster-slurm-example-hpc-workload/dagster_slurm_example_hpc_workload/ray/
+
+# Run with default settings
+python process_documents_docling.py
+
+# Or customize parameters
+python process_documents_docling.py \
+  --input-glob "data/**/*.pdf" \
+  --output-dir "out/my_docs" \
+  --num-workers 4 \
+  --batch-size 8
+```
+
+**What happens:**
+
+1. Script runs as a normal Python program on your local machine
+2. Initializes a local Ray instance automatically
+3. Processes documents in parallel using available CPU cores
+4. Converts PDFs to markdown using docling
+5. Prints results to console (no Dagster, no metadata tracking)
+
+**Best for:**
+
+- Quick testing and experimentation
+- Debugging conversion issues
+- Learning how the processing works
[…]

[…]
