+++ projects/dagster-slurm/dagster_slurm/resources/compute.py
@@ -761,7 +761,6 @@ class ComputeResource(ConfigurableResource):
             env_overrides = {}
             if self.pack_platform:
                 env_overrides["SLURM_PACK_PLATFORM"] = self.pack_platform
-            logger.debug(f"Pack command (one line): {shlex.join(pack_cmd)}")
             pack_file = pack_environment_with_pixi(
                 pack_cmd=pack_cmd, env_overrides=env_overrides or None
             )
