+++ projects/dagster-slurm/dagster_slurm/pipes_clients/slurm_pipes_client.py
@@ -556,7 +556,6 @@ class SlurmPipesClient(PipesClient):
         env_overrides = {}
         if self.pack_platform:
             env_overrides["SLURM_PACK_PLATFORM"] = self.pack_platform
-        self.logger.debug(f"Pack command (one line): {shlex.join(pack_cmd)}")
         pack_file = pack_environment_with_pixi(
             pack_cmd=pack_cmd, env_overrides=env_overrides or None
         )
