+++ projects/dagster-slurm/dagster_slurm/resources/session.py
@@ -246,6 +246,9 @@ class SlurmSessionResource(ConfigurableResource):
         job_id = int(match.group(1))
         self.logger.info(f"Allocation submitted: job {job_id}")
 
+        # Query estimated start time for pending allocation
+        self._log_estimated_start_time(job_id)
+
         # Wait for allocation to start
         self._wait_for_allocation_start(job_id, working_dir, timeout=120)
 
@@ -261,6 +264,31 @@ class SlurmSessionResource(ConfigurableResource):
             working_dir=working_dir,
             config=self,
         )
+
+    def _log_estimated_start_time(self, job_id: int) -> None:
+        """Query and log the estimated start time for a pending allocation job."""
+        try:
+            # Query estimated start time
+            output = self._ssh_pool.run(  # type: ignore
+                f"squeue --start -j {job_id} -h -o '%S' 2>/dev/null || true"
+            )
+            estimated_start = output.strip()
+
+            if estimated_start and estimated_start != "N/A":
+                self.logger.info(
+                    f"Allocation job {job_id} --> Estimated start time: {estimated_start}"
+                )
+            else:
+                # No estimate available (job may start immediately or already started)
+                self.logger.debug(
+                    f"No estimated start time for allocation {job_id} (immediate start or already running)"
+                )
+
+        except Exception as e:
+            # Non-critical - just log as debug and continue
+            self.logger.debug(
+                f"Could not query estimated start time for allocation {job_id}: {e}"
+            )
 
     def _wait_for_allocation_start(
         self,
