+++ examples/projects/dagster-slurm-example-shared/dagster_slurm_example_shared/__init__.py
@@ -1,0 +1,19 @@
+"""Shared utilities and configurations for dagster-slurm examples."""
+
+from dagster_slurm_example_shared.paths import (
+    SITE_CONFIGS,
+    get_base_data_path,
+    get_base_input_path,
+    get_base_output_path,
+    get_scratch_path,
+    get_site_info,
+)
+
+__all__ = [
+    "get_base_data_path",
+    "get_base_input_path",
+    "get_base_output_path",
+    "get_scratch_path",
+    "get_site_info",
+    "SITE_CONFIGS",
+]
