+++ projects/dagster-slurm/dagster_slurm/pipes_clients/slurm_pipes_client.py
@@ -57,7 +57,6 @@ class SlurmPipesClient(PipesClient):
         pack_platform: Optional[str] = None,
         pre_deployed_env_path: Optional[str] = None,
         cache_inject_globs: Optional[list[str]] = None,
-        cancel_on_interrupt: bool = False,
     ):
         """Args:
         slurm_resource: Slurm cluster configuration
@@ -71,9 +70,6 @@ class SlurmPipesClient(PipesClient):
             should affect the environment cache key. If None, all --inject patterns
             from the pack command are hashed. Use this to exclude workload-specific
             packages from cache invalidation (e.g., only include base libraries).
-        cancel_on_interrupt: If True, cancel SLURM jobs when Dagster is interrupted (Ctrl+C, restart).
-            If False, jobs continue running independently. Useful for long-running jobs that should
-            survive Dagster restarts. Default: False (jobs survive restarts).
 
         """
         super().__init__()
@@ -84,7 +80,6 @@ class SlurmPipesClient(PipesClient):
         self.debug_mode = debug_mode
         self.auto_detect_platform = auto_detect_platform
         self.pack_platform = pack_platform
-        self.cancel_on_interrupt = cancel_on_interrupt
         self.logger = get_dagster_logger()
         self.metrics_collector = SlurmMetricsCollector()
         self._control_path = None
@@ -155,16 +150,12 @@ class SlurmPipesClient(PipesClient):
 
         # Setup cancellation handler
         def handle_cancellation(signum, frame):
-            self.logger.warning(f"⚠️  Cancellation signal received (signal {signum})")
-            self._cancellation_requested = True
+            self.logger.warning(f"⚠️  Termination signal received (signal {signum})")
+
             if self._current_job_id:
-                if self.cancel_on_interrupt:
-                    self._cancel_slurm_job(self._current_job_id)
-                else:
-                    self.logger.info(
-                        f"Job {self._current_job_id} will continue running (cancel_on_interrupt=False). "
-                        f"To cancel manually: scancel {self._current_job_id}"
-                    )
+                self.logger.info(f"Cancelling SLURM job {self._current_job_id}...")
+                self._cancellation_requested = True
+                self._cancel_slurm_job(self._current_job_id)
 
         # Register signal handlers
         original_sigint = signal.signal(signal.SIGINT, handle_cancellation)
@@ -274,6 +265,7 @@ class SlurmPipesClient(PipesClient):
                             ssh_pool=ssh_pool,
                             message_reader=message_reader,
                             extra_slurm_opts=extra_slurm_opts,
+                            op_context=context.op_execution_context,
                         )
 
                     self._maybe_emit_final_logs(
@@ -347,18 +339,12 @@ class SlurmPipesClient(PipesClient):
                         self.logger.warning(f"Cleanup failed: {e}")
 
         except Exception as e:
-            # Cancel job if it's running (unless cancel_on_interrupt=False)
+            # Cancel job if it's running and not already cancelled
             if self._current_job_id and not self._cancellation_requested:
-                if self.cancel_on_interrupt:
-                    self.logger.warning(
-                        f"Cancelling job {self._current_job_id} due to error"
-                    )
-                    self._cancel_slurm_job(self._current_job_id)
-                else:
-                    self.logger.warning(
-                        f"Job {self._current_job_id} will continue running despite error "
-                        f"(cancel_on_interrupt=False). Monitor with: squeue -j {self._current_job_id}"
-                    )
+                self.logger.warning(
+                    f"Cancelling job {self._current_job_id} due to error"
+                )
+                self._cancel_slurm_job(self._current_job_id)
 
             self.logger.error(f"Execution failed: {e}")
 
[…]
