### 2026-01-23 3a925b3
fix: datalab + UX refinements
