+++ examples/projects/dagster-slurm-example/dagster_slurm_example/defs/ray/docling_assets.py
@@ -6,66 +6,26 @@ on Slurm clusters, it spawns a multi-node Ray cluster for parallel processing.
 """
 
 import os
-from pathlib import Path
 
 import dagster as dg
 from dagster_slurm import ComputeResource, RayLauncher, SlurmRunConfig
-from dagster_slurm_example_shared import get_base_output_path
-from pydantic import Field
-
-
-def _default_docling_glob() -> str | None:
-    """Return default input glob based on deployment environment.
-
-    - Local/development: Use local example data directory
-    - Docker SLURM: Use Docker container data path
-    - Supercomputer: None (must be configured via launchpad)
-    """
-    deployment = os.getenv("DAGSTER_DEPLOYMENT", "development")
-
-    # Docker SLURM deployments use container path
-    # Use *.pdf to match files in the directory root (not just subdirectories)
-    if "docker" in deployment:
-        return "/home/submitter/dagster-slurm-data/*.pdf"
-
-    # Supercomputer: no default path (user must configure via launchpad)
-    if "supercomputer" in deployment:
-        return None
-
-    # For local development, use local example data
-    examples_dir = Path(__file__).resolve().parents[5]
-    data_dir = examples_dir / "data"
-    if data_dir.exists() and any(data_dir.rglob("*.pdf")):
-        return str(data_dir / "**" / "*.pdf")
-
-    # Unknown deployment: no default
-    return None
-
-
-def _default_docling_output_dir() -> str:
-    """Return default output directory based on deployment environment.
-
-    Uses shared path configuration from dagster_slurm_example_shared.
-    - Local/development: /tmp/dagster_output/docling
-    - Docker: $DATA/output/docling
-    - musica/vsc5: $DATA/output/docling
-    - datalab: $HOME/output/docling
-    """
-    base_output = get_base_output_path()
-    return f"{base_output}/docling"
+from pydantic import Field, model_validator
 
 
 class DoclingRunConfig(SlurmRunConfig):
     """Configuration for docling document processing."""
 
-    input_glob: str | None = Field(
-        default_factory=_default_docling_glob,
-        description="Glob pattern for input PDF files to process (relative to project root or absolute path). "
-        "Required for supercomputer deployments - configure via launchpad.",
+    input_glob: str = Field(
+        description="Glob pattern for input PDF files to process (absolute path). "
+        "Examples:\n"
+        "  - Docker: /home/submitter/dagster-slurm-data/*.pdf\n"
+        "  - Supercomputer: /scratch/username/data/*.pdf\n"
+        "  - Development: /path/to/your/data/**/*.pdf",
     )
     output_dir: str = Field(
-        default_factory=_default_docling_output_dir,
-        description="Directory where processed documents will be saved",
+        default="$HOME/output/docling",
+        description="Directory where processed documents will be saved. "
+        "Default: $HOME/output/docling (expanded on compute node)",
     )
     num_workers: int = Field(
         default=2,
@@ -80,18 +40,32 @@ class DoclingRunConfig(SlurmRunConfig):
         description="Number of documents to process per batch",
     )
 
+    @model_validator(mode="after")
+    def validate_input_glob_not_empty(self):
+        """Validate that input_glob is not empty."""
+        if not self.input_glob or not self.input_glob.strip():
+            raise ValueError(
+                "input_glob is required and cannot be empty. "
+                "Please provide a glob pattern in the Dagster launchpad, "
+                "e.g., '/scratch/username/data/*.pdf' or '/home/user/documents/**/*.pdf'"
+            )
+        return self
+
 
 class LargeScaleDoclingRunConfig(SlurmRunConfig):
     """Configuration for large-scale multi-node docling processing."""
 
[…]
