+++ projects/dagster-slurm/dagster_slurm/launchers/ray.py
@@ -268,16 +268,27 @@ class RayLauncher(ComputeLauncher):
     temp_dir_arg=""
     RAY_TMP_DIR=""  # Global for cleanup function
     if [[ -n "${{SLURM_JOB_ID:-}}" ]]; then
-      # Use SLURM_TMPDIR if available (per-job temp dir), otherwise /tmp with job ID
+      # Use SLURM_TMPDIR if available (per-job temp dir)
       if [[ -n "${{SLURM_TMPDIR:-}}" ]]; then
         RAY_TMP_DIR="${{SLURM_TMPDIR}}/ray"
+        mkdir -p "$RAY_TMP_DIR"
+        echo "[$({date_fmt})] Using SLURM_TMPDIR: $RAY_TMP_DIR"
       else
-        RAY_TMP_DIR="/tmp/ray_job_${{SLURM_JOB_ID}}"
+        # Try very short /tmp path first (avoids Unix socket 107-byte path limit)
+        SHORT_TMP="/tmp/r${{SLURM_JOB_ID}}"
+        if mkdir -p "$SHORT_TMP" 2>/dev/null; then
+          RAY_TMP_DIR="$SHORT_TMP"
+          echo "[$({date_fmt})] Using /tmp: $RAY_TMP_DIR"
+        else
+          # /tmp not writable - use $HOME/.r<jobid> (much shorter than working dir)
+          RAY_TMP_DIR="$HOME/.r${{SLURM_JOB_ID}}"
+          mkdir -p "$RAY_TMP_DIR"
+          echo "[$({date_fmt})] Using HOME temp dir: $RAY_TMP_DIR"
+        fi
       fi
-      mkdir -p "$RAY_TMP_DIR"
       export RAY_TMPDIR="$RAY_TMP_DIR"
       temp_dir_arg="--temp-dir=$RAY_TMP_DIR"
-      echo "[$({date_fmt})] Ray temp directory: $RAY_TMP_DIR"
+      echo "[$({date_fmt})] Ray temp directory: $RAY_TMP_DIR ($(echo -n "$RAY_TMP_DIR" | wc -c) chars)"
     fi
 {override_block}    # Start local Ray cluster
     echo "[$({date_fmt})] Starting local Ray cluster"
@@ -396,7 +407,6 @@ class RayLauncher(ComputeLauncher):
         Generates a robust Ray cluster startup script with proper shutdown.
         """
         redis_pw = self.redis_password or "$(uuidgen)"
-        temp_dir_path = "/tmp/ray-$SLURM_JOB_ID"
 
         common_args = []
         if self.object_store_memory_gb is not None:
@@ -427,7 +437,7 @@ class RayLauncher(ComputeLauncher):
                 "--dashboard-port=$dash_port",
                 f"--num-gpus={self.num_gpus_per_node}",
                 "--redis-password=$redis_password",
-                f"--temp-dir={temp_dir_path}",
+                f"--temp-dir=$RAY_CLUSTER_TMP",
             ]
             + common_args
             + self.ray_start_args
@@ -438,6 +448,7 @@ class RayLauncher(ComputeLauncher):
             "--address=$ip_head",
             "--redis-password=$redis_password",
             f"--num-gpus={self.num_gpus_per_node}",
+            "--temp-dir=$RAY_CLUSTER_TMP",
         ] + common_args
 
         head_cmd_str = " \\\n    ".join(head_args)
@@ -477,7 +488,7 @@ class RayLauncher(ComputeLauncher):
 
         # Final sync of worker logs
         worker_logs_dir="{working_dir}/ray_logs/worker_$(hostname)"
[…]
