### 2026-01-23 d2120fb
fix: ray distributed mode
