+++ projects/dagster-slurm/dagster_slurm/launchers/ray.py
@@ -58,10 +58,12 @@ class RayLauncher(ComputeLauncher):
         description="Optional shell commands to run before ray start (e.g., ulimit).",
     )
     worker_cpu_bind: Optional[str] = Field(
-        default=None,
+        default="none",
         description=(
-            "Optional value for srun --cpu-bind when starting workers. "
-            "Leave unset to inherit Slurm defaults."
+            "Value for srun --cpu-bind when starting workers. "
+            "Default 'none' disables CPU binding to avoid conflicts with job allocation. "
+            "Other values: 'cores', 'threads', 'sockets', etc. "
+            "Set to Python None (not the string) to omit the flag entirely."
         ),
     )
     use_head_ip: bool = Field(
@@ -263,8 +265,8 @@ class RayLauncher(ComputeLauncher):
     fi
     head_adv="$head_bind_addr"
     if [[ "$head_adv" == *:* ]]; then head_adv="[$head_adv]"; fi
-    # Use /tmp for Ray sockets (Unix socket path limit: 107 bytes)
-    # $HOME/dagster_runs paths are too long and cause socket errors
+    # Use short path for Ray sockets (Unix socket path limit: 107 bytes)
+    # Must be node-local storage - Unix sockets don't work over NFS
     temp_dir_arg=""
     RAY_TMP_DIR=""  # Global for cleanup function
     if [[ -n "${{SLURM_JOB_ID:-}}" ]]; then
@@ -272,19 +274,17 @@ class RayLauncher(ComputeLauncher):
       if [[ -n "${{SLURM_TMPDIR:-}}" ]]; then
         RAY_TMP_DIR="${{SLURM_TMPDIR}}/ray"
         mkdir -p "$RAY_TMP_DIR"
-        echo "[$({date_fmt})] Using SLURM_TMPDIR: $RAY_TMP_DIR"
+        echo "[$({date_fmt})] Using SLURM_TMPDIR: $RAY_TMP_DIR (node-local)"
+      elif mkdir -p "/tmp/r${{SLURM_JOB_ID}}" 2>/dev/null; then
+        RAY_TMP_DIR="/tmp/r${{SLURM_JOB_ID}}"
+        echo "[$({date_fmt})] Using /tmp: $RAY_TMP_DIR (node-local)"
+      elif mkdir -p "/var/tmp/r${{SLURM_JOB_ID}}" 2>/dev/null; then
+        RAY_TMP_DIR="/var/tmp/r${{SLURM_JOB_ID}}"
+        echo "[$({date_fmt})] Using /var/tmp: $RAY_TMP_DIR (node-local)"
       else
-        # Try very short /tmp path first (avoids Unix socket 107-byte path limit)
-        SHORT_TMP="/tmp/r${{SLURM_JOB_ID}}"
-        if mkdir -p "$SHORT_TMP" 2>/dev/null; then
-          RAY_TMP_DIR="$SHORT_TMP"
-          echo "[$({date_fmt})] Using /tmp: $RAY_TMP_DIR"
-        else
-          # /tmp not writable - use $HOME/.r<jobid> (much shorter than working dir)
-          RAY_TMP_DIR="$HOME/.r${{SLURM_JOB_ID}}"
-          mkdir -p "$RAY_TMP_DIR"
-          echo "[$({date_fmt})] Using HOME temp dir: $RAY_TMP_DIR"
-        fi
+        RAY_TMP_DIR="$HOME/.r${{SLURM_JOB_ID}}"
+        mkdir -p "$RAY_TMP_DIR"
[…]
