+++ .github/workflows/library.yaml
@@ -215,6 +215,9 @@ jobs:
           SLURM_EDGE_NODE_USER: submitter
           SLURM_EDGE_NODE_PASSWORD: submitter
           SLURM_DEPLOYMENT_BASE_PATH: /home/submitter/pipelines/deployments
+          # Set SLURM_JOB_ID to enable Ray port randomization in local mode tests
+          # This prevents port 6379 conflicts (Redis default port) in CI
+          SLURM_JOB_ID: ${{ github.run_id }}
         run: pixi run -e build --frozen testpixi-integration
 
       - name: Stop SLURM Cluster and Clean Up
