### 2026-01-23 39a9cc7
feat: example docling ray (#90)
* add docling document processing example
* streamline development experience
* first steps towards implementing ste-ssh OIDC flows
