### 2026-01-23 291d8f5
1.10.0-rc.1 [skip ci]
Automatically generated by python-semantic-release
