### 2026-02-16 #9795 @jan-janssen
@HPicatto Can you clarify the contributions of the individual authors? Based on the contributions recorded by Github, https://github.com/ascii-supply-networks/dagster-slurm/graphs/contributors , the `dagster-slurm` package seems to be primarily developed by Georg Heiler with 72 commits while you have 4 commits and Maximilian Hess has 2 commits. Why is Martin Pfister not included in the list of authors when he also has 2 commits?
