### 2026-02-16 #9795 @jan-janssen
@HPicatto While the documentation covers how to setup the tooling for development, the release process and how to test with a local SLURM Docker container, the documentation does not cover how you want others to contribute: https://dagster-slurm.geoheil.com/docs/contributing/

Please address the following three points:
1. Contribute to the software
2. Report issues or problems with the software
3. Seek support
