### 2026-02-16 #9795 @jan-janssen
@HPicatto your statement "Existing solutions either ignore the orchestrator (hand-written Slurm scripts) or bypass the scheduler entirely (running ad-hoc services). This leads to duplicated logic, fragile deployments, and a loss of telemetry." is wrong. A large number of Python based task schedulers have been developed in the last decade, some examples include:
* parsl as introduced in https://dl.acm.org/doi/10.1145/3307681.3325400
* jobflow in https://doi.org/10.21105/joss.05995
* executorlib in https://doi.org/10.21105/joss.07782 (the one I am personally involved in.)

All three use Python decorators similar to the `@dg.asset()` decorator from dragster,  track dependencies of tasks, handle SSH connections to remote computing resources and so on. Furthermore, a more extensive list of workflow managers is available at https://github.com/common-workflow-language/common-workflow-language/wiki/Existing-Workflow-systems still not all of these provide Python interfaces.

So while I agree with the statement that the dragster Data Orchestrator Platform does not have a SLURM interface, it is unclear to me why you had to reinvent the wheel in terms of interfacing to SLURM, remote transfer and dependency management rather than building on top of one existing solution.

Furthermore, there are existing interfaces which provide an abstraction layer over different job schedulers, like https://doi.org/10.1109/e-Science58273.2023.10254912 which you could have used as well.

Please update the **statement of need** how `dagster-slurm` compares to existing workflow solutions commonly used in the Python community and why existing interfaces for the SLRUM job scheduler were insufficient.
