### 2026-02-16 #9795 @jan-janssen
@HPicatto Please be careful with mixing the current implementation with planned developments. Statements like "Spark and custom launchers are planned for near future release" should be exclusively in the `Impact and Future Work` section, not the `System Overview` section.

As Dagster identifies itself as Data Orchestrator Platform, it is unclear how the data transfer is implemented in `dagster-slurm`. In the example https://dagster-slurm.geoheil.com/docs/applications/document-preprocessing-docling#test-data the data is still copied on the command line using ssh. Does `dagster-slurm` provide internal tools for copying the data? I see the use of the `SLURM_DEPLOYMENT_BASE_PATH` variable to transfer the python environment but it seems to require the corresponding data to be available on the HPC already. In the example the local folder `data` is used.

In the context of Dagster, does the `dagster-slurm` interface implement the required `io managers` for transferring data to the HPC and back from the HPC https://docs.dagster.io/api/dagster/io-managers#dagster.IOManager ? Or is the current implementation of `dagster-slurm` limited to the `ConfigurableResource` interface which manages the computing resources but does not provide the file IO, meaning the user has to manually transfer input files and output files when using the `dagster-slurm` interface, in contrast to other Dagster interfaces.
