+++ docs/paper.bib
@@ -170,12 +170,63 @@ abstract = {Reproducibility is widely acknowledged as a fundamental principle in
   url          = {https://doi.org/10.5281/zenodo.15306610},
 }
 @misc{eisty2025,
-      title={Ten Essential Guidelines for Building High-Quality Research Software}, 
+      title={Ten Essential Guidelines for Building High-Quality Research Software},
       author={Nasir U. Eisty and David E. Bernholdt and Alex Koufos and David J. Luet and Miranda Mundt},
       year={2025},
       eprint={2507.16166},
       archivePrefix={arXiv},
       primaryClass={cs.SE},
-      url={https://arxiv.org/abs/2507.16166}, 
+      url={https://arxiv.org/abs/2507.16166},
       doi={10.48550/arXiv.2507.16166},
+}
+
+@inproceedings{parsl,
+  author    = {Babuji, Yadu and Woodard, Anna and Li, Zhuozhao and Katz, Daniel S. and Clifford, Ben and Kumar, Rohan and Lacinski, Lukasz and Chard, Ryan and Wozniak, Justin M. and Foster, Ian and Wilde, Michael and Chard, Kyle},
+  title     = {Parsl: Pervasive Parallel Programming in Python},
+  booktitle = {Proceedings of the 28th International Symposium on High-Performance Parallel and Distributed Computing},
+  series    = {HPDC '19},
+  pages     = {25--36},
+  year      = {2019},
+  publisher = {ACM},
+  doi       = {10.1145/3307681.3325400},
+}
+
+@article{executorlib,
+  author  = {Janssen, Jan and Taylor, Michael Gilbert and Yang, Ping and Neugebauer, Joerg and Perez, Danny},
+  title   = {executorlib -- Up-scaling Python workflows for hierarchical heterogeneous high-performance computing},
+  journal = {Journal of Open Source Software},
+  volume  = {10},
+  number  = {108},
+  pages   = {7782},
+  year    = {2025},
+  doi     = {10.21105/joss.07782},
+}
+
+@article{jobflow,
+  author  = {Rosen, Andrew S. and Gallant, Max and George, Janine and Riebesell, Janosh and Saez, Hrushikesh and Ganose, Alex M. and Jain, Anubhav},
+  title   = {jobflow: Computational Workflows Made Simple},
+  journal = {Journal of Open Source Software},
+  volume  = {9},
+  number  = {93},
+  pages   = {5995},
+  year    = {2024},
+  doi     = {10.21105/joss.05995},
+}
+
+@inproceedings{psij,
+  author    = {Hategan-Marandiuc, Mihael and Merzky, Andre and Collier, Nicholson and Maheshwari, Ketan and Ozik, Jonathan and Turilli, Matteo and Wilke, Andreas and Wozniak, Justin M. and Chard, Kyle and Foster, Ian and Ferreira da Silva, Rafael and Jha, Shantenu and Laney, Daniel},
+  title     = {{PSI/J}: A Portable Interface for Submitting, Monitoring, and Managing Jobs},
+  booktitle = {2023 IEEE 19th International Conference on e-Science (e-Science)},
+  pages     = {1--10},
+  year      = {2023},
+  publisher = {IEEE},
+  doi       = {10.1109/e-Science58273.2023.10254912},
+}
+
+@misc{cwl_wiki,
+  author = {{Common Workflow Language Community}},
+  title  = {Existing Workflow Systems},
+  year   = {2024},
+  url    = {https://s.apache.org/existing-workflow-systems},
+  note   = {Accessed 17 February 2026},
 }
