### 2026-02-17 #9795 @geoHeil
> [@HPicatto](https://github.com/HPicatto) Can you clarify the contributions of the individual authors? Based on the contributions recorded by Github, https://github.com/ascii-supply-networks/dagster-slurm/graphs/contributors , the `dagster-slurm` package seems to be primarily developed by Georg Heiler with 72 commits while you have 4 commits and Maximilian Hess has 2 commits. Why is Martin Pfister not included in the list of authors when he also has 2 commits?

Martin was the supervisor during the hackathon and contributed a very small feature. His Organiaztion was already acknowledge. He initially could not actively participate in writing the publication due to resource constraints. But this changed recently (see also the result of many discussions enabling step-ssh auth integration)- so I added him now. I hope this better reflects the contributon
