+++ projects/dagster-slurm/dagster_slurm_test/test_cancellation.py
@@ -183,27 +183,67 @@ def test_pre_submission_check_blocks_on_sigterm():
 # ---------------------------------------------------------------------------
 
 
+def _submit_sleep_job(
+    ssh_pool, run_dir: str, job_name: str, sleep_seconds: int = 30
+) -> int:
+    """Submit a simple bash sleep job and return the Slurm job ID."""
+    import re
+    import tempfile
+    from pathlib import Path
+
+    ssh_pool.run(f"mkdir -p {run_dir}")
+
+    job_script = f"#!/bin/bash\necho started\nsleep {sleep_seconds}\necho done\n"
+    script_path = f"{run_dir}/job.sh"
+
+    with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".sh") as f:
+        f.write(job_script)
+        local_temp = f.name
+    ssh_pool.upload_file(local_temp, script_path)
+    Path(local_temp).unlink()
+    ssh_pool.run(f"chmod +x {script_path}")
+
+    output = ssh_pool.run(
+        f"sbatch -J {job_name} -D {run_dir} "
+        f"-o {run_dir}/slurm-%j.out -e {run_dir}/slurm-%j.err "
+        f"-t 00:05:00 -c 1 {script_path}"
+    )
+    match = re.search(r"Submitted batch job (\d+)", output)
+    assert match, f"Could not parse job ID from: {output}"
+    return int(match.group(1))
+
+
+def _wait_for_running(ssh_pool, job_id: int, timeout: int = 30) -> str:
+    """Poll until the job reaches RUNNING (or a terminal state). Return state."""
+    import time
+
+    for _ in range(timeout):
+        out = ssh_pool.run(f"squeue -h -j {job_id} -o '%T' 2>/dev/null || true")
+        state = out.strip()
+        if state == "RUNNING":
+            return state
+        if state in {"COMPLETED", "FAILED", "CANCELLED", "TIMEOUT"}:
+            return state
+        time.sleep(1)
+    return state
+
+
+def _get_final_state(ssh_pool, job_id: int) -> str:
+    """Return the final state of a completed job via sacct."""
+    out = ssh_pool.run(f"sacct -X -n -j {job_id} -o State 2>/dev/null || true")
+    return out.strip().split()[0] if out.strip() else ""
+
+
 @pytest.mark.needs_slurm_docker
 @pytest.mark.slow
 def test_terminate_run_cancels_slurm_job(
     slurm_resource_for_testing,
     slurm_cluster_ready,
-    temp_dir,
 ):
-    """Submit a long-running job, set is_interrupt_requested, verify CANCELLED."""
-    import re
+    """Submit a long-running job, call _cancel_slurm_job, verify CANCELLED in sacct."""
     import time
 
     from dagster_slurm.helpers.ssh_pool import SSHConnectionPool
-
-    # Create a long-running test payload
-    payload = temp_dir / "long_running.py"
-    payload.write_text(
-        "import time\n"
-        "print('started', flush=True)\n"
-        "time.sleep(30)\n"
-        "print('done', flush=True)\n"
-    )
 
     client = SlurmPipesClient(
         slurm_resource=slurm_resource_for_testing,
@@ -212,52 +252,25 @@ def test_terminate_run_cancels_slurm_job(
 
     ssh_pool = SSHConnectionPool(slurm_resource_for_testing.ssh)
     with ssh_pool:
-        remote_base = "/home/submitter/dagster_ci_runs"
-        run_dir = f"{remote_base}/test_cancel_{int(time.time())}"
-        ssh_pool.run(f"mkdir -p {run_dir}")
[…]
