### 2026-02-17 42ffd7c
fix: mark as not slow
