+++ projects/dagster-slurm/dagster_slurm/pipes_clients/slurm_pipes_client.py
@@ -14,6 +14,8 @@ from typing import Any, Dict, Optional
 
 from dagster import (
     AssetExecutionContext,
+    DagsterRunStatus,
+    OpExecutionContext,
     PipesClient,
     PipesEnvContextInjector,
     get_dagster_logger,
@@ -172,6 +174,86 @@ class SlurmPipesClient(PipesClient):
                 self._control_path = ssh_pool.control_path
 
                 remote_base = self._get_remote_base(run_id, ssh_pool)
+
+                # --- Reattach path ---
+                # Check if a previous run left a Slurm job (active or
+                # already completed) for this asset.  If so, skip env
+                # prep / payload upload / submission.
+                previous = self._find_previous_slurm_job(context, ssh_pool, remote_base)
+                if previous:
+                    job_id, run_dir, job_status = previous
+                    self.logger.info(
+                        f"Reattaching to Slurm job {job_id} "
+                        f"(status={job_status}) in {run_dir}"
+                    )
+
+                    messages_path = f"{run_dir}/messages.jsonl"
+                    context_injector = PipesEnvContextInjector()
+                    message_reader = SSHMessageReader(
+                        remote_path=messages_path,
+                        ssh_config=self.slurm.ssh,
+                        control_path=ssh_pool.control_path,
+                        ssh_pool=ssh_pool,
+                    )
+
+                    with open_pipes_session(
+                        context=context.op_execution_context,
+                        context_injector=context_injector,
+                        message_reader=message_reader,
+                        extras=extras,
+                    ) as session:
+                        self._current_job_id = job_id
+
+                        if job_status == "active":
+                            # Job is still running — monitor until completion
+                            self._wait_for_job_with_streaming(
+                                job_id,
+                                ssh_pool,
+                                run_dir,
+                                message_reader=message_reader,
+                                op_context=context.op_execution_context,
+                            )
+
+                        # Job completed (either just now or before we got here)
+                        self._maybe_emit_final_logs(
+                            message_reader=message_reader,
+                            ssh_pool=ssh_pool,
+                            run_dir=run_dir,
+                            job_id=job_id,
+                        )
+
+                        self.logger.info(
+                            f"Reattached job {job_id} completed successfully"
+                        )
+
+                        # Collect metrics
+                        try:
+                            metrics = self.metrics_collector.collect_job_metrics(
+                                job_id, ssh_pool
+                            )
+                            metadata = {
+                                "slurm_job_id": job_id,
+                                "node_hours": metrics.node_hours,
+                                "cpu_efficiency_pct": round(
+                                    metrics.cpu_efficiency * 100, 2
+                                ),
+                                "max_memory_mb": round(metrics.max_rss_mb, 2),
+                                "elapsed_seconds": round(metrics.elapsed_seconds, 2),
+                            }
+                            context.add_output_metadata(metadata)
+                        except Exception as e:
+                            self.logger.warning(f"Failed to collect metrics: {e}")
+
+                    # Cleanup (unless debug mode)
+                    if not self.debug_mode:
+                        try:
+                            self._schedule_async_cleanup(ssh_pool, run_dir)
+                        except Exception as e:
+                            self.logger.warning(f"Cleanup failed: {e}")
+
+                    return PipesClientCompletedInvocation(session)
+
[…]
