+++ projects/dagster-slurm/dagster_slurm_test/test_cancellation.py
@@ -125,28 +125,88 @@ def test_exception_handler_cancels_on_real_error():
 
 
 def test_dagster_ui_cancel_still_works():
-    """When op_context.is_interrupt_requested is True, scancel IS called."""
+    """When SIGTERM + run is CANCELING, scancel IS called.
+
+    Dagster UI "Terminate Run" sets the run status to CANCELING in the
+    database, then sends SIGTERM.  The polling loop queries the database
+    and should cancel the Slurm job.
+    """
     client = _make_client()
     client._current_job_id = 67890
     client._ssh_pool = MagicMock()
-    client._sigterm_received = False
+    client._sigterm_received = True  # SIGTERM received
     client._cancellation_requested = False
 
-    mock_op_context = MagicMock()
-    mock_op_context.is_interrupt_requested = True
-
-    with patch.object(client, "_cancel_slurm_job") as mock_cancel:
-        # Simulate the polling loop's Dagster UI cancellation check
-        if (
-            mock_op_context
-            and hasattr(mock_op_context, "is_interrupt_requested")
-            and mock_op_context.is_interrupt_requested
-        ):
-            client._cancellation_requested = True
-            client._cancel_slurm_job(client._current_job_id)
+    with (
+        patch.object(client, "_cancel_slurm_job") as mock_cancel,
+        patch.object(client, "_is_run_canceling", return_value=True),
+    ):
+        # Simulate the polling loop's SIGTERM + CANCELING path
+        if client._sigterm_received:
+            if client._is_run_canceling(None):
+                client._cancellation_requested = True
+                client._cancel_slurm_job(client._current_job_id)
 
         mock_cancel.assert_called_once_with(67890)
         assert client._cancellation_requested is True
+
+
+def test_sigterm_without_canceling_status_does_not_cancel():
+    """SIGTERM without CANCELING run status = reboot, should NOT cancel."""
+    client = _make_client()
+    client._current_job_id = 67890
+    client._ssh_pool = MagicMock()
+    client._sigterm_received = True  # SIGTERM received
+    client._cancellation_requested = False
+
+    with (
+        patch.object(client, "_cancel_slurm_job") as mock_cancel,
+        patch.object(client, "_is_run_canceling", return_value=False),
+    ):
+        # Simulate the polling loop's SIGTERM path (reboot scenario)
+        if client._sigterm_received:
+            if client._is_run_canceling(None):
+                client._cancellation_requested = True
+                client._cancel_slurm_job(client._current_job_id)
+
+        mock_cancel.assert_not_called()
+        assert client._cancellation_requested is False
+
+
+def test_is_run_canceling_returns_true_for_canceling_run():
+    """_is_run_canceling returns True when run status is CANCELING."""
+    from dagster import DagsterRunStatus
+
+    client = _make_client()
+    mock_run = MagicMock()
+    mock_run.status = DagsterRunStatus.CANCELING
+
+    mock_op_context = MagicMock()
+    mock_op_context.run_id = "test-run-id"
+    mock_op_context.instance.get_run_by_id.return_value = mock_run
+
+    assert client._is_run_canceling(mock_op_context) is True
+
+
+def test_is_run_canceling_returns_false_for_started_run():
+    """_is_run_canceling returns False when run status is STARTED (reboot)."""
+    from dagster import DagsterRunStatus
+
+    client = _make_client()
+    mock_run = MagicMock()
+    mock_run.status = DagsterRunStatus.STARTED
+
[…]
