### 2026-02-17 3f14afe
fix: mostly there now
