+++ projects/dagster-slurm/dagster_slurm_test/test_cancellation.py
@@ -223,269 +223,41 @@ def test_sigterm_received_reset_in_finally():
     assert client._sigterm_received is False
 
 
-def test_pre_submission_check_blocks_on_sigterm():
-    """Pre-submission checks should abort when _sigterm_received is True."""
+def test_pre_submission_check_allows_sigterm_on_reboot():
+    """Pre-submission check should NOT abort on SIGTERM when the run is NOT
+    CANCELING (i.e. Dagster reboot, not user cancel)."""
     client = _make_client()
     client._sigterm_received = True
     client._cancellation_requested = False
 
-    # Simulate the pre-submission guard
-    if client._cancellation_requested or client._sigterm_received:
-        raised = True
-    else:
-        raised = False
+    with patch.object(client, "_is_run_canceling", return_value=False):
+        # Simulate the pre-submission guard (matches production code)
+        if client._cancellation_requested or (
+            client._sigterm_received and client._is_run_canceling(None)
+        ):
+            raised = True
+        else:
+            raised = False
+
+    assert raised is False
+
+
+def test_pre_submission_check_blocks_on_sigterm_with_canceling():
+    """Pre-submission check SHOULD abort when SIGTERM + run is CANCELING
+    (user clicked Terminate Run)."""
+    client = _make_client()
+    client._sigterm_received = True
+    client._cancellation_requested = False
+
+    with patch.object(client, "_is_run_canceling", return_value=True):
+        if client._cancellation_requested or (
+            client._sigterm_received and client._is_run_canceling(None)
+        ):
+            raised = True
+        else:
+            raised = False
 
     assert raised is True
-
-
-# ---------------------------------------------------------------------------
-# Reattach tests (no Slurm docker needed)
-# ---------------------------------------------------------------------------
-
-
-def test_find_previous_slurm_job_active_from_parent():
-    """When a parent run's dir has slurm output and squeue shows RUNNING."""
-    client = _make_client()
-
-    mock_context = MagicMock()
-    mock_context.run.parent_run_id = "parent-run-123"
-    mock_context.run.run_id = "current-run-456"
-    mock_context.run.job_name = "my_job"
-    mock_context.instance.get_runs.return_value = []
-
-    mock_ssh_pool = MagicMock()
-    mock_ssh_pool.run.side_effect = [
-        # ls slurm-*.out → found a file
-        "/remote/pipelines/runs/parent-run-123/slurm-99.out\n",
-        # squeue -j 99 → RUNNING
-        "RUNNING\n",
-    ]
-
-    result = client._find_previous_slurm_job(
-        mock_context, mock_ssh_pool, "/remote/pipelines"
-    )
-
-    assert result is not None
-    job_id, run_dir, status = result
-    assert job_id == 99
-    assert run_dir == "/remote/pipelines/runs/parent-run-123"
-    assert status == "active"
-
-
-def test_find_previous_slurm_job_completed_via_sacct():
-    """When squeue is empty but sacct shows COMPLETED, return completed job."""
-    client = _make_client()
-
-    mock_context = MagicMock()
-    mock_context.run.parent_run_id = "parent-run-123"
-    mock_context.run.run_id = "current-run-456"
-    mock_context.run.job_name = "my_job"
-    mock_context.instance.get_runs.return_value = []
-
-    mock_ssh_pool = MagicMock()
-    mock_ssh_pool.run.side_effect = [
-        # ls slurm-*.out → found a file
-        "/remote/pipelines/runs/parent-run-123/slurm-99.out\n",
[…]

[…]
