+++ projects/dagster-slurm/dagster_slurm_test/test_cancellation.py
@@ -5,6 +5,7 @@ Unit tests verify that:
 - The exception handler skips scancel when SIGTERM was received
 - The exception handler still calls scancel on genuine errors
 - Dagster UI "Terminate Run" still calls scancel via the polling loop
+- Reattach helpers work correctly (tags, asset key extraction, job lookup)
 """
 
 import signal
@@ -18,7 +19,12 @@ from dagster_slurm import (
     SSHConnectionResource,
     SlurmQueueConfig,
 )
-from dagster_slurm.pipes_clients.slurm_pipes_client import SlurmPipesClient
+from dagster_slurm.pipes_clients.slurm_pipes_client import (
+    SlurmPipesClient,
+    _TAG_ASSET_KEY,
+    _TAG_JOB_ID,
+    _TAG_RUN_DIR,
+)
 
 
 def _make_client() -> SlurmPipesClient:
@@ -258,6 +264,252 @@ def test_pre_submission_check_blocks_on_sigterm_with_canceling():
             raised = False
 
     assert raised is True
+
+
+# ---------------------------------------------------------------------------
+# Reattach helper unit tests
+# ---------------------------------------------------------------------------
+
+
+def test_store_job_tags():
+    """_store_job_tags writes correct tag keys/values to the instance."""
+    client = _make_client()
+    mock_op_context = MagicMock()
+    mock_op_context.run_id = "test-run-id"
+
+    client._store_job_tags(mock_op_context, 42, "/tmp/run_dir", "my_asset")
+
+    mock_op_context.instance.add_run_tags.assert_called_once_with(
+        "test-run-id",
+        {
+            _TAG_JOB_ID: "42",
+            _TAG_RUN_DIR: "/tmp/run_dir",
+            _TAG_ASSET_KEY: "my_asset",
+        },
+    )
+
+
+def test_store_job_tags_no_asset_key():
+    """_store_job_tags omits asset key tag when asset_key is None."""
+    client = _make_client()
+    mock_op_context = MagicMock()
+    mock_op_context.run_id = "test-run-id"
+
+    client._store_job_tags(mock_op_context, 42, "/tmp/run_dir", None)
+
+    mock_op_context.instance.add_run_tags.assert_called_once_with(
+        "test-run-id",
+        {
+            _TAG_JOB_ID: "42",
+            _TAG_RUN_DIR: "/tmp/run_dir",
+        },
+    )
+
+
+def test_store_job_tags_none_context():
+    """_store_job_tags does nothing when op_context is None."""
+    client = _make_client()
+    # Should not raise
+    client._store_job_tags(None, 42, "/tmp/run_dir", "my_asset")
+
+
+def test_get_asset_key_string_single_asset():
+    """_get_asset_key_string returns the asset key as a string."""
+    client = _make_client()
+    mock_op_context = MagicMock()
+    mock_op_context.asset_key = MagicMock()
+    mock_op_context.asset_key.__str__ = MagicMock(return_value="AssetKey(['my_asset'])")
+
+    result = client._get_asset_key_string(mock_op_context)
+    assert result == "AssetKey(['my_asset'])"
+
+
+def test_get_asset_key_string_no_asset():
+    """_get_asset_key_string returns None for non-asset ops."""
+    client = _make_client()
+    mock_op_context = MagicMock(spec=[])  # No attributes at all
+
+    result = client._get_asset_key_string(mock_op_context)
+    assert result is None
+
+
+def test_find_reattachable_job_from_parent_run():
+    """_find_reattachable_job finds job from parent run's tags (retry path)."""
+    client = _make_client()
+    mock_op_context = MagicMock()
+    mock_op_context.dagster_run.parent_run_id = "parent-run-123"
+
+    parent_run = MagicMock()
+    parent_run.tags = {
[…]
