+++ projects/dagster-slurm/dagster_slurm/pipes_clients/slurm_pipes_client.py
@@ -22,6 +22,7 @@ from dagster import (
     open_pipes_session,
 )
 from dagster._core.pipes.client import PipesClientCompletedInvocation
+from dagster._utils.interrupts import _received_interrupt
 
 from ..helpers.env_packaging import compute_env_cache_key, pack_environment_with_pixi
 from ..helpers.message_readers import SSHMessageReader
@@ -31,6 +32,10 @@ from ..helpers.ssh_pool import SSHConnectionPool
 from ..launchers.base import ComputeLauncher
 from ..resources.session import SlurmSessionResource
 from ..resources.slurm import SlurmResource
+
+_TAG_JOB_ID = "dagster_slurm/job_id"
+_TAG_RUN_DIR = "dagster_slurm/run_dir"
+_TAG_ASSET_KEY = "dagster_slurm/asset_key"
 
 
 class SlurmPipesClient(PipesClient):
@@ -163,6 +168,11 @@ class SlurmPipesClient(PipesClient):
                 "and will continue running on the cluster."
             )
             self._sigterm_received = True
+            # Prevent Dagster's executor from detecting the interrupt and
+            # failing the run after our step completes successfully.  This
+            # is needed for the in-process executor where our handler and
+            # Dagster's capture_interrupts share the same process.
+            _received_interrupt["received"] = False
 
         # Register signal handlers
         original_sigint = signal.signal(signal.SIGINT, handle_signal)
@@ -179,6 +189,137 @@ class SlurmPipesClient(PipesClient):
                 messages_path = f"{run_dir}/messages.jsonl"
                 self.logger.debug(f"Creating remote run directory: {run_dir}")
                 ssh_pool.run(f"mkdir -p {run_dir}")
+
+                # --- Reattach path: check for a previous Slurm job ---
+                # On retry after a failed run (e.g. Dagster restart with
+                # multiprocess executor), we look for the Slurm job that
+                # was submitted by the previous attempt.  Three outcomes:
+                #   1. Job still running → reattach and monitor
+                #   2. Job already COMPLETED → collect metrics, skip submission
+                #   3. Job failed / not found → fall through to normal path
+                op_ctx = context.op_execution_context
+                asset_key_str = self._get_asset_key_string(op_ctx)
+                reattach_info = self._find_reattachable_job(
+                    op_ctx, ssh_pool, asset_key_str
+                )
+                if reattach_info:
+                    old_job_id_str = reattach_info["job_id"]
+                    old_run_dir = reattach_info["run_dir"]
+                    old_job_id = int(old_job_id_str)
+                    old_job_state = self._get_job_state(old_job_id, ssh_pool).upper()
+
+                    if self._is_job_still_running(old_job_id, ssh_pool):
+                        # Case 1: job still active — reattach to monitoring
+                        self.logger.info(
+                            f"Reattaching to running Slurm job {old_job_id} "
+                            f"(state={old_job_state}, run_dir={old_run_dir})"
+                        )
+                        self._current_job_id = old_job_id
+                        job_id = old_job_id
+                        run_dir = old_run_dir
+                        self._store_job_tags(op_ctx, old_job_id, run_dir, asset_key_str)
+
+                        message_reader = SSHMessageReader(
+                            remote_path=f"{run_dir}/messages.jsonl",
+                            ssh_config=self.slurm.ssh,
+                            control_path=ssh_pool.control_path,
+                            ssh_pool=ssh_pool,
+                        )
+                        context_injector = PipesEnvContextInjector()
+                        with open_pipes_session(
+                            context=op_ctx,
+                            context_injector=context_injector,
+                            message_reader=message_reader,
+                            extras=extras,
+                        ) as session:
+                            self._wait_for_job_with_streaming(
+                                old_job_id,
[…]

[…]
