+++ examples/projects/dagster-slurm-example-hpc-workload/dagster_slurm_example_hpc_workload/ray/process_documents_docling.py
@@ -495,3 +495,7 @@ if __name__ == "__main__":
         finally:
             if ray.is_initialized():
                 ray.shutdown()
+
+
+def foo():
+    return 1
