+++ pyproject.toml
@@ -19,7 +19,7 @@ build-backend = "uv_build"
 
 [dependency-groups]
 dev = [
-  "pytest>=8.3.4,<9",
+  "pytest>=8.3.4,<10",
   "pytest-mock>=3.14.0,<4",
   "pytest-cov~=7.0.0",
   "ruff>=0.12.8,<1",
@@ -43,7 +43,7 @@ dprint = ">=0.50.0,<1"
 uv = ">=0.8.6"
 build = ">=1.3.0"
 dagster-slurm = { path = "./projects/dagster-slurm", editable = true }
-pytest = ">=8.3.4,<9"
+pytest = ">=8.3.4,<10"
 pytest-cov = "~=7.0.0"
 pexpect = "~=4.9.0"
 ty = ">=0.0.4,<1"
