+++ docs/paper.md
@@ -22,11 +22,15 @@ authors:
   - name: Georg Heiler
     orcid: 0000-0002-8684-1163
     affiliation: "1, 2"
+  - name: Martin Pfister
+    affiliation: "3"
 affiliations:
  - name: Complexity Science Hub Vienna (CSH)
    index: 1
  - name: Austrian Supply Chain Intelligence Institute (ASCII)
    index: 2
+ - name: AI Factory Austria
+   index: 3
 
 date: 1st November 2025
 bibliography: paper.bib
@@ -46,8 +50,8 @@ This paper introduces **dagster-slurm**, an open-source integration that allows 
 The project packages dependencies with Pixi @pixi, submits workloads through Slurm using Dagster Pipes @dagsterpipes, and streams logs plus scheduler metrics back to the Dagster UI.
 
 The key contribution is a unified compute resource (`ComputeResource`) that hides SSH transport (including password-only jump hosts and OTP prompts), dependency packaging, and queue configuration while still respecting Slurm’s scheduling semantics.
-Today the project ships with two production-ready execution modes—`local` for laptop/CI development and `slurm` for one-job-per-asset submissions.
-Experimental support for session-based reuse and heterogeneous jobs is work in progress. Launchers for Bash, Ray, and Spark (work in progress) ship out of the box and can be extended.
+The project ships two production-ready execution modes—`local` for laptop/CI development and `slurm` for one-job-per-asset submissions—and two stable launchers: Bash for script-based workloads and Ray for multi-node distributed computing.
+Experimental support for Spark, session-based allocation reuse, and heterogeneous jobs is under active development.
 
 # Statement of Need
 
@@ -57,13 +61,30 @@ Reproducibility challenges in HPC environments are well documented, and Antunes 
 Courtes et al. @Courtes9882991 further examine the tension between reproducibility and performance, demonstrating that these goals need not be mutually exclusive.
 
 The 2024 Community Workshop on Practical Reproducibility in HPC produced a comprehensive report highlighting cost-effective reproducibility challenges and the need for tools that bridge development and production environments @keahey_2025_15306610.
-Existing solutions either ignore the orchestrator (hand-written Slurm scripts) or bypass the scheduler entirely (running ad-hoc services).
-This leads to duplicated logic, fragile deployments, and a loss of telemetry. **dagster-slurm** was created to:
+A rich ecosystem of Python-based HPC workflow tools already exists.
+Parsl @parsl provides a parallel scripting library with multi-site execution, dependency tracking, and SSH-based remote submission.
+Executorlib @executorlib extends Python's standard `concurrent.futures.Executor` to HPC schedulers, offering per-function resource control, local-to-cluster portability, and support for heterogeneous workloads combining MPI and GPU libraries.
+Jobflow @jobflow manages complex computational workflows with decorator-based task definitions, and PSI/J @psij offers a portable submission interface across schedulers.
+The Common Workflow Language community maintains an extensive catalogue of further systems @cwl_wiki.
+These frameworks provide mature solutions for constructing and executing task graphs on HPC resources, often including local execution modes for rapid prototyping.
 
-- Preserve Dagster’s asset-based design, lineage tracking, and alerting for workloads that ultimately run on Slurm-managed hardware.
-- Remove the need to rewrite orchestration glue when moving from development to production supercomputers.
-- Provide a batteries-included path for packaging Python environments reproducibly (Pixi https://pixi.sh/latest/ + pixi-pack, https://github.com/Quantco/pixi-pack) and deploying them in air-gapped environments.
-- Offer a clear path to advanced HPC patterns (session reuse for long-lived clusters, heterogeneous Slurm jobs) while keeping the current stable surface area intentionally small; these features are being iterated on in the open.
+dagster-slurm does not aim to replace any of these tools; it occupies a complementary niche.
[…]
