+++ docs/docs/api/api_core.md
@@ -20,7 +20,7 @@ Run Dagster assets on Slurm clusters with support for:
 
 ### *class* dagster_slurm.BashLauncher(\*\*data)
 
-Bases: [`ComputeLauncher`](#dagster_slurm.ComputeLauncher)
+Bases: [`ComputeLauncher`](#class-dagster_slurmcomputelauncherdata)
 
 Executes Python scripts via bash.
 
@@ -137,7 +137,7 @@ slurm=slurm,
 
 #### debug_mode *: `bool`*
 
-#### default_launcher *: `Annotated`[`Union`[[`ComputeLauncher`](#dagster_slurm.ComputeLauncher), `PartialResource`]]*
+#### default_launcher *: `Annotated`[`Union`[[`ComputeLauncher`](#class-dagster_slurmcomputelauncherdata), `PartialResource`]]*
 
 #### default_skip_payload_upload *: `bool`*
 
@@ -149,7 +149,7 @@ Get appropriate Pipes client for this mode.
 
 - **Parameters:**
   - **context** (`InitResourceContext`) – Dagster resource context
-  - **launcher** (`Optional`[[`ComputeLauncher`](#dagster_slurm.ComputeLauncher)]) – Override launcher (uses default if None)
+  - **launcher** (`Optional`[[`ComputeLauncher`](#class-dagster_slurmcomputelauncherdata)]) – Override launcher (uses default if None)
 - **Returns:**
   LocalPipesClient or SlurmPipesClient
 
@@ -189,7 +189,7 @@ Execute asset with optional resource overrides.
 - **Parameters:**
   - **context** – Dagster execution context
   - **payload_path** (`str`) – Path to Python script
-  - **launcher** (`Optional`[[`ComputeLauncher`](#dagster_slurm.ComputeLauncher)]) – Override launcher for this asset
+  - **launcher** (`Optional`[[`ComputeLauncher`](#class-dagster_slurmcomputelauncherdata)]) – Override launcher for this asset
   - **extra_slurm_opts** (`Optional`[`Dict`[`str`, `Any`]]) – Override Slurm options (non-session mode)
     - nodes: int
     - cpus_per_task: int
@@ -208,7 +208,7 @@ Execute asset with optional resource overrides.
     or asset metadata key ‘skip_slurm_payload_upload’.
   - **remote_payload_path** (`Optional`[`str`]) – Remote path to an existing payload when skipping upload.
     Falls back to config.remote_payload_path or asset metadata.
-  - **config** (`Optional`[[`SlurmRunConfig`](#dagster_slurm.SlurmRunConfig)]) – Optional SlurmRunConfig for run-time configuration via launchpad.
+  - **config** (`Optional`[[`SlurmRunConfig`](#class-dagster_slurmslurmrunconfigconfig_dict)]) – Optional SlurmRunConfig for run-time configuration via launchpad.
     Values from config are used as defaults, but explicit parameters take precedence.
   - **\*\*kwargs** – Passed to client.run()
 - **Yields:**
@@ -281,7 +281,7 @@ Only waits in queue ONCE, but each asset gets the resources it needs.
     > - mem: str (default: “4G”)
     > - gpus_per_node: int (default: 0)
     > - time_limit: str (default: “01:00:00”)
-  - **launchers** (`Optional`[`Dict`[`str`, [`ComputeLauncher`](#dagster_slurm.ComputeLauncher)]]) – Optional dict mapping asset_key to ComputeLauncher
[…]
