### 2026-02-18 51134a3
1.10.0-rc.2 [skip ci]
Automatically generated by python-semantic-release
