+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 <!-- version list -->
 
+## v1.10.0-rc.2 (2026-02-18)
+
+### Bug Fixes
+
+- Address reviewers comments ([#97](https://github.com/ascii-supply-networks/dagster-slurm/pull/97),
+  [`530e8ab`](https://github.com/ascii-supply-networks/dagster-slurm/commit/530e8ab4a52ff69bd257f4a48686cf80282bac33))
+
+- Revision 1 ([#97](https://github.com/ascii-supply-networks/dagster-slurm/pull/97),
+  [`530e8ab`](https://github.com/ascii-supply-networks/dagster-slurm/commit/530e8ab4a52ff69bd257f4a48686cf80282bac33))
+
+
 ## v1.10.0-rc.1 (2026-01-23)
 
 ### Features
