### 2026-02-19 #9795 @HPicatto
@jan-janssen the pull request that @geoHeil mention was merged if you please review the new document. I think we make clear what is implemented, the current state of workflow managers, and why dagster-slurm is different.
