### 2026-02-19 7c1d3dd
feat: cancellation AND continued remote execution when dagster is restarted (#98)
* fix: refine cancellation vs automatic continuation for long running jobs where dagster is restarted
