+++ projects/dagster-slurm/dagster_slurm/pipes_clients/slurm_pipes_client.py
@@ -14,12 +14,15 @@ from typing import Any, Dict, Optional
 
 from dagster import (
     AssetExecutionContext,
+    DagsterRunStatus,
+    OpExecutionContext,
     PipesClient,
     PipesEnvContextInjector,
     get_dagster_logger,
     open_pipes_session,
 )
 from dagster._core.pipes.client import PipesClientCompletedInvocation
+from dagster._utils.interrupts import _received_interrupt
 
 from ..helpers.env_packaging import compute_env_cache_key, pack_environment_with_pixi
 from ..helpers.message_readers import SSHMessageReader
@@ -29,6 +32,10 @@ from ..helpers.ssh_pool import SSHConnectionPool
 from ..launchers.base import ComputeLauncher
 from ..resources.session import SlurmSessionResource
 from ..resources.slurm import SlurmResource
+
+_TAG_JOB_ID = "dagster_slurm/job_id"
+_TAG_RUN_DIR = "dagster_slurm/run_dir"
+_TAG_ASSET_KEY = "dagster_slurm/asset_key"
 
 
 class SlurmPipesClient(PipesClient):
@@ -86,6 +93,7 @@ class SlurmPipesClient(PipesClient):
         self._current_job_id = None
         self._ssh_pool = None
         self._cancellation_requested = False
+        self._sigterm_received = False
         self.pre_deployed_env_path = pre_deployed_env_path
         self.cache_inject_globs = cache_inject_globs
 
@@ -148,18 +156,27 @@ class SlurmPipesClient(PipesClient):
         run_dir = None
         job_id = None
 
-        # Setup cancellation handler
-        def handle_cancellation(signum, frame):
-            self.logger.warning(f"⚠️  Termination signal received (signal {signum})")
-
-            if self._current_job_id:
-                self.logger.info(f"Cancelling SLURM job {self._current_job_id}...")
-                self._cancellation_requested = True
-                self._cancel_slurm_job(self._current_job_id)
+        # Setup signal handler – on SIGTERM/SIGINT we only set a flag so the
+        # polling loop can exit cleanly.  We intentionally do NOT call scancel
+        # here because a Dagster UI reboot sends SIGTERM and the Slurm job
+        # should keep running on the cluster.  User-initiated cancellation
+        # (via the "Terminate Run" button) is handled by the polling loop's
+        # `is_interrupt_requested` check which *does* call scancel.
+        def handle_signal(signum, frame):
+            self.logger.warning(
+                f"Signal {signum} received. Slurm job will NOT be cancelled "
+                "and will continue running on the cluster."
+            )
+            self._sigterm_received = True
+            # Prevent Dagster's executor from detecting the interrupt and
+            # failing the run after our step completes successfully.  This
+            # is needed for the in-process executor where our handler and
+            # Dagster's capture_interrupts share the same process.
+            _received_interrupt["received"] = False
 
         # Register signal handlers
-        original_sigint = signal.signal(signal.SIGINT, handle_cancellation)
-        original_sigterm = signal.signal(signal.SIGTERM, handle_cancellation)
+        original_sigint = signal.signal(signal.SIGINT, handle_signal)
+        original_sigterm = signal.signal(signal.SIGTERM, handle_signal)
 
         try:
             with ssh_pool:
@@ -167,10 +184,142 @@ class SlurmPipesClient(PipesClient):
                 self._control_path = ssh_pool.control_path
 
                 remote_base = self._get_remote_base(run_id, ssh_pool)
+
                 run_dir = f"{remote_base}/runs/{run_id}"
                 messages_path = f"{run_dir}/messages.jsonl"
                 self.logger.debug(f"Creating remote run directory: {run_dir}")
                 ssh_pool.run(f"mkdir -p {run_dir}")
+
[…]
