+++ projects/dagster-slurm/dagster_slurm_test/test_dagster_reattach.py
@@ -1,0 +1,743 @@
+"""Dagster-level integration tests for SIGTERM behaviour and reattach.
+
+These tests use ``dagster.materialize()`` with real DagsterInstance /
+AssetExecutionContext.  Only the SSH layer (``SSHConnectionPool``) is
+mocked so no Docker cluster is needed.
+
+Crucially, ``_execute_standalone`` and ``_wait_for_job_with_streaming``
+are NOT mocked – the real code paths execute so that the polling loop's
+SIGTERM handling is fully exercised.
+"""
+
+import threading
+from unittest.mock import MagicMock, patch
+
+from dagster import (
+    AssetExecutionContext,
+    DagsterInstance,
+    asset,
+    materialize,
+)
+
+from dagster_slurm import (
+    BashLauncher,
+    ComputeResource,
+    SlurmQueueConfig,
+    SlurmResource,
+    SSHConnectionResource,
+)
+from dagster_slurm.config.environment import ExecutionMode
+from dagster_slurm.pipes_clients.slurm_pipes_client import (
+    SlurmPipesClient,
+    _TAG_JOB_ID,
+    _TAG_RUN_DIR,
+)
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+
+def _make_slurm_compute() -> ComputeResource:
+    """ComputeResource in SLURM mode with a dummy SSH target."""
+    ssh = SSHConnectionResource(
+        host="localhost",
+        port=2223,
+        user="testuser",
+        password="testpass",
+    )
+    slurm = SlurmResource(
+        ssh=ssh,
+        queue=SlurmQueueConfig(
+            partition="test",
+            time_limit="00:10:00",
+            cpus=2,
+            mem="1G",
+        ),
+        remote_base="/tmp/dagster_test",
+    )
+    return ComputeResource(
+        mode=ExecutionMode.SLURM,
+        default_launcher=BashLauncher(),
+        slurm=slurm,
+        cleanup_on_failure=False,
+    )
+
+
+def _mock_ssh_pool():
+    """Create a mock SSHConnectionPool that behaves as a context manager."""
+    pool = MagicMock()
+    pool.__enter__ = MagicMock(return_value=pool)
+    pool.__exit__ = MagicMock(return_value=False)
+    pool.control_path = "/tmp/ctrl"
+    return pool
+
+
+class _SshRouter:
+    """Route ``ssh_pool.run(cmd)`` calls to appropriate fake responses.
+
+    This lets the real ``_execute_standalone`` and
+    ``_wait_for_job_with_streaming`` run against fake SSH output.
+    """
+
+    def __init__(self):
+        self.submitted_job_id = 42
+        # Track poll count to fire SIGTERM at the right time
+        self._poll_count = 0
+        self._sigterm_target: SlurmPipesClient | None = None
+        self._sigterm_after_polls = 1  # fire SIGTERM after first poll
+        # After this many polls the job transitions to COMPLETED
+        self._complete_after_polls = 3
+
+    def __call__(self, cmd: str) -> str:
+        # Remote base / home detection
+        if cmd.startswith("echo $HOME"):
+            return "/tmp/dagster_test"
+
+        # mkdir
+        if cmd.startswith("mkdir"):
+            return ""
+
+        # File upload verification (test -f ... && wc -l)
+        if "test -f" in cmd and "wc -l" in cmd:
+            return "10 /tmp/job.sh"
+
+        # chmod
+        if cmd.startswith("chmod"):
+            return ""
+
+        # sbatch submission
+        if "sbatch" in cmd:
+            return f"Submitted batch job {self.submitted_job_id}"
+
+        # squeue poll (job state check)
[…]
