### 2026-02-19 93e6a2f
feat: cancellation AND continued remote execution when dagster is restarted (#98)
* fix: refine cancellation vs automatic continuation for long running jobs where dagster is restarted
