+++ docs/docs/how-to/troubleshooting.md
@@ -39,3 +39,109 @@ find /home/$USER/dagster_runs/runs -mindepth 1 -maxdepth 1 -mtime +3 -exec rm -r
 ```
 
 The `cleanup_on_failure` flag continues to apply—failed runs respect `debug_mode` and skip removal.
+
+## Laptop and one-off researcher deployments
+
+Not every HPC workflow needs a persistent Dagster server. A common pattern is running Dagster directly on a researcher's laptop or workstation: submit long-running Slurm jobs, close the lid, and check results hours or days later. `dagster-slurm` is designed to support this workflow.
+
+**Why this works without persistent infrastructure:**
+
+- **Slurm is the durable layer.** Once `sbatch` submits a job, the HPC scheduler owns it. The job runs to completion regardless of what happens to the Dagster process that submitted it. Logs, outputs, and exit codes are stored on the cluster.
+- **Dagster is the orchestration layer.** It packages the environment, submits the job, and streams logs while it is running. But it does not need to be running for the Slurm job to finish.
+- **Reattach on restart.** When you reopen your laptop and start Dagster again, `dagster-slurm` detects the still-running (or already-completed) Slurm job and picks up where it left off. It **never resubmits** — the retry only monitors or collects results from the existing job.
+
+**Recommended setup for laptop deployments:**
+
+```python
+from dagster import Definitions, define_asset_job, in_process_executor
+
+# In-process executor keeps runs green through Dagster restarts.
+# For laptop use this is almost always the right choice — there is
+# no benefit to multiprocess when you are submitting to Slurm anyway.
+slurm_job = define_asset_job(
+    "slurm_job",
+    executor_def=in_process_executor,
+)
+```
+
+With this setup, you can:
+
+1. **Start Dagster** and materialize assets — jobs are submitted to the cluster.
+2. **Stop Dagster** (or close your laptop) — Slurm jobs continue running on the cluster.
+3. **Restart Dagster** — if the process was still monitoring a job when it was stopped, the in-process executor catches the signal and the run completes on the next startup. If the process was already gone, a retry reattaches.
+
+There is no need for a persistent Dagster daemon, database server, or always-on VM. The HPC cluster provides the durability; Dagster provides the developer experience (log streaming, metadata, lineage, retries).
+
+:::tip When to deploy persistent Dagster infrastructure
+
+Consider a dedicated Dagster server (e.g. on a VM or Kubernetes) when you need:
+
+- **Scheduled runs** (cron-like triggers via `ScheduleDefinition`)
+- **Sensors** that react to external events
+- **Team-wide visibility** across multiple users' runs
+- **Automated backfill orchestration** across many partitions
+
+For a single researcher iterating on HPC workloads, a laptop deployment is simpler and sufficient.
+
+:::
+
+## Dagster restarts and Slurm job survival
+
+When Dagster restarts (e.g. a deploy, code location reload, or process crash), it sends `SIGTERM` to running processes. How this affects your Slurm jobs depends on the **executor** you are using.
+
+### In-process executor (recommended for Slurm workloads)
+
+With the in-process executor, `dagster-slurm` catches the `SIGTERM`, lets the Slurm job continue running, and keeps monitoring it. **The run completes successfully** — the Slurm job is never cancelled.
+
+This is the recommended configuration when you expect frequent Dagster restarts and want green run results:
+
+```python
+from dagster import Definitions, define_asset_job, in_process_executor
+
+slurm_job = define_asset_job(
+    "slurm_job",
+    executor_def=in_process_executor,
+)
+```
+
+### Multiprocess executor
+
+:::warning Multiprocess executor always fails runs on SIGTERM
+
[…]
