### 2026-02-22 7ea16fa
fix: update deps for demo
