+++ examples/pyproject.toml
@@ -52,6 +52,7 @@ pexpect = "~=4.9.0"
 [tool.pixi.feature.my-infrastructure.pypi-dependencies]
 dagster-slurm-example = { path = "./projects/dagster-slurm-example", editable = true }
 dagster-ray = ">=0.4.0,<1"
+metaxy = { version = ">=0.1.0", extras = ["duckdb", "dagster"] }
 
 [tool.pixi.feature.submission.dependencies]
 [tool.pixi.feature.submission.pypi-dependencies]
@@ -64,6 +65,7 @@ ray = { version = "~=2.53.0", extras = ["data", "default"] }
 dagster-ray = ">=0.4.0,<0.25"
 dagster-pipes = "~=1.12.0"
 dagster = "~=1.12.0"
+metaxy = { version = ">=0.1.0", extras = ["delta", "ray"] }
 
 [tool.pixi.feature.workload-document-processing.dependencies]
 pylatexenc = ">=2.10,<3"
