### 2026-02-22 08b92ce
feat: first step towards skills here
