+++ SKILLS.md
@@ -1,0 +1,23 @@
+# Skills Index
+
+This repository includes agent skills via the submodule:
+
+- `.agent-skills/dagster-skills`
+
+Initialize after clone:
+
+```bash
+git submodule update --init --recursive
+```
+
+## Available Skills
+
+- `dagster-expert`: Dagster architecture, definitions, assets, automation, CLI workflows.
+- `dagster-integrations`: Guide to Dagster integration libraries and tool selection.
+- `dignified-python`: Production Python coding standards and modern typing/style patterns.
+
+## Skill Paths
+
+- `.agent-skills/dagster-skills/skills/dagster-expert`
+- `.agent-skills/dagster-skills/skills/dagster-integrations`
+- `.agent-skills/dagster-skills/skills/dignified-python`
