+++ examples/projects/dagster-slurm-example-shared/dagster_slurm_example_shared/__init__.py
@@ -1,5 +1,13 @@
 """Shared utilities and configurations for dagster-slurm examples."""
 
+from dagster_slurm_example_shared.metaxy_features import (
+    ConvertedDocuments,
+    Embeddings,
+    InputTexts,
+    ProcessedNumbers,
+    RawNumbers,
+    SourceDocuments,
+)
 from dagster_slurm_example_shared.paths import (
     SITE_CONFIGS,
     get_base_data_path,
@@ -16,4 +24,10 @@ __all__ = [
     "get_scratch_path",
     "get_site_info",
     "SITE_CONFIGS",
+    "ConvertedDocuments",
+    "Embeddings",
+    "InputTexts",
+    "ProcessedNumbers",
+    "RawNumbers",
+    "SourceDocuments",
 ]
