+++ SKILLS.md
@@ -15,9 +15,11 @@ git submodule update --init --recursive
 - `dagster-expert`: Dagster architecture, definitions, assets, automation, CLI workflows.
 - `dagster-integrations`: Guide to Dagster integration libraries and tool selection.
 - `dignified-python`: Production Python coding standards and modern typing/style patterns.
+- `metaxy`: Metaxy feature definitions, versioning, metadata stores, CLI, and testing patterns.
 
 ## Skill Paths
 
 - `.agent-skills/dagster-skills/skills/dagster-expert`
 - `.agent-skills/dagster-skills/skills/dagster-integrations`
 - `.agent-skills/dagster-skills/skills/dignified-python`
+- `.agent-skills/metaxy/SKILL.md`
