### 2026-02-22 d3c0747
feat: mini ray example
