### 2026-02-22 57313b9
feat: docling example complete (for local execution)
