+++ examples/projects/dagster-slurm-example-hpc-workload/dagster_slurm_example_hpc_workload/ray/process_documents_docling_metaxy.py
@@ -1,0 +1,434 @@
+"""HPC workload: Document processing with docling + metaxy incremental tracking.
+
+This script runs on the HPC cluster and:
+1. Discovers input files via glob and registers them in metaxy
+2. Uses store.resolve_update() to find new/changed documents
+3. Processes only those documents with Ray Data + docling
+4. Writes results via MetaxyDatasink
+
+Reuses BasicDocumentConverter and warmup_model_cache from the existing
+docling workload (process_documents_docling.py).
+"""
+
+import glob
+import hashlib
+import os
+import time
+from pathlib import Path
+from typing import Any
+
+import metaxy as mx
+
+# metaxy.toml is uploaded as sibling of this script via extra_files
+os.environ.setdefault("METAXY_CONFIG", str(Path(__file__).parent / "metaxy.toml"))
+# Import shared feature definitions to register them with metaxy
+import dagster_slurm_example_shared.metaxy_features  # noqa: F401
+import polars as pl
+import pyarrow as pa
+import pyarrow.compute as pc
+import ray
+import ray.data as rd
+from dagster_pipes import DagsterPipesError, PipesContext, open_dagster_pipes
+from dagster_slurm_example_shared import get_base_output_path
+from metaxy.ext.ray import MetaxyDatasink
+
+from dagster_slurm_example_hpc_workload.ray.process_documents_docling import (
+    BasicDocumentConverter,
+    warmup_model_cache,
+)
+
+METAXY_TIMESTAMP_COLUMNS = (
+    "metaxy_created_at",
+    "metaxy_updated_at",
+    "metaxy_deleted_at",
+)
+
+
+def _coerce_metaxy_timestamps_to_utc(table: pa.Table) -> pa.Table:
+    for name in METAXY_TIMESTAMP_COLUMNS:
+        idx = table.schema.get_field_index(name)
+        if idx < 0:
+            continue
+        field = table.schema.field(idx)
+        if not pa.types.is_timestamp(field.type):
+            continue
+        if field.type.tz == "UTC":
+            continue
+        table = table.set_column(
+            idx,
+            name,
+            pc.cast(table.column(idx), pa.timestamp(field.type.unit, tz="UTC")),
+        )
+    return table
+
+
+def _upsert_column(table: pa.Table, name: str, array: pa.Array) -> pa.Table:
+    idx = table.schema.get_field_index(name)
+    if idx >= 0:
+        return table.set_column(idx, name, array)
+    return table.append_column(name, array)
+
+
+class MetaxyDoclingMapper:
+    """Ray map_batches callable that converts documents and produces metaxy-compatible output."""
+
+    def __init__(self, output_dir: str):
+        self.output_dir = output_dir
+        self.converter = BasicDocumentConverter()
+
+    def __call__(self, batch: pa.Table) -> pa.Table:  # noqa: C901
+        from docling.datamodel.document import ConversionStatus
+        from docling_core.types.doc import ImageRefMode
+
+        source_paths = batch.column(
+            batch.schema.get_field_index("source_path")
+        ).to_pylist()
+        markdown_paths: list[str] = []
+        num_pages_list: list[int] = []
+        status_list: list[str] = []
+        elapsed_list: list[float] = []
+
+        for source_path in source_paths:
+            t0 = time.time()
+            try:
+                result = self.converter.convert_one(source_path)
+
+                if result.status != ConversionStatus.SUCCESS:
+                    elapsed = time.time() - t0
+                    markdown_paths.append("")
+                    num_pages_list.append(0)
+                    status_list.append(f"failed: {result.status}")
+                    elapsed_list.append(round(elapsed, 2))
+                    continue
+
+                document = result.document
+                pages = document.num_pages()
+                doc_name = Path(source_path).stem
[…]
