### 2026-02-22 #9795 @junghans
> **Paper file info**:
>
> ⚠️ Wordcount for `paper.md` is **1853**

@geoHeil, this is nearly double the recommended max length (1000 words), can you try to shorten the paper a bit?
