+++ examples/projects/dagster-slurm-example/dagster_slurm_example/defs/metaxy/simple_metaxy_assets.py
@@ -50,6 +50,46 @@ def _compute_results(df: pl.DataFrame) -> pl.DataFrame:
     )
 
 
+def _is_schema_mismatch_error(exc: Exception) -> bool:
+    error_text = str(exc)
+    return (
+        "Column name/value mismatch for insert" in error_text
+        or "expected" in error_text
+        and "columns but" in error_text
+        and "values were supplied" in error_text
+    )
+
+
+def _ensure_current_feature_schema(
+    store: mx.MetadataStore,
+    feature_key: str,
+    required_output_columns: set[str],
+    context_logger: dg.DagsterLogManager | None = None,
+) -> None:
+    """One-time migration helper for legacy feature tables.
+
+    Older local duckdb files can contain feature tables created with a previous
+    schema. If required output columns are missing, drop feature metadata and let
+    Metaxy recreate it on next write.
+    """
+    try:
+        with store:
+            schema_names = set(store.read(feature_key).collect_schema().names())
+    except Exception:
+        # Feature not materialized yet (or no readable metadata) - nothing to migrate.
+        return
+
+    if required_output_columns.issubset(schema_names):
+        return
+
+    logger = context_logger or dg.get_dagster_logger()
+    logger.info(
+        f"Detected legacy schema for {feature_key}; applying one-time metadata reset."
+    )
+    with store:
+        store.drop_feature_metadata(feature_key)
+
+
 @mxd.metaxify
 @dg.asset(
     metadata={"metaxy/feature": "example/raw_numbers"},
@@ -107,6 +147,12 @@ def processed_numbers(  # noqa: C901
 
     Only these samples are processed, skipping already-processed unchanged ones.
     """
+    _ensure_current_feature_schema(
+        store_simple,
+        "example/processed_numbers",
+        {"result", "value_bucket"},
+    )
+
     with store_simple:
         increment = store_simple.resolve_update("example/processed_numbers")
 
@@ -140,11 +186,30 @@ def processed_numbers(  # noqa: C901
         f"{len(to_process)} samples (incremental)"
     )
 
-    with store_simple:
-        if len(new_with_results) > 0:
-            store_simple.write("example/processed_numbers", new_with_results)
-        if len(stale_with_results) > 0:
-            store_simple.write("example/processed_numbers", stale_with_results)
+    try:
+        with store_simple:
+            if len(new_with_results) > 0:
+                store_simple.write("example/processed_numbers", new_with_results)
+            if len(stale_with_results) > 0:
+                store_simple.write("example/processed_numbers", stale_with_results)
+    except Exception as e:
+        if not _is_schema_mismatch_error(e):
+            raise
+
+        logger = dg.get_dagster_logger()
+        logger.warning(
+            "Detected legacy schema mismatch in example/processed_numbers. "
+            "Dropping feature metadata and retrying with current schema."
+        )
+        with store_simple:
+            store_simple.drop_feature_metadata("example/processed_numbers")
+            retry_increment = store_simple.resolve_update("example/processed_numbers")
+            retry_new = _compute_results(retry_increment.new.to_polars())
+            retry_stale = _compute_results(retry_increment.stale.to_polars())
+            if len(retry_new) > 0:
+                store_simple.write("example/processed_numbers", retry_new)
+            if len(retry_stale) > 0:
+                store_simple.write("example/processed_numbers", retry_stale)
 
     return to_process.select(
         pl.col("sample_uid"), pl.col("result"), pl.col("value_bucket")
@@ -169,6 +234,13 @@ def partitioned_processed_numbers(
 
     partition_filter = _partition_pl_filter(partition_key)
     target_filter = _partition_nw_filter(partition_key)
+
+    _ensure_current_feature_schema(
[…]
