+++ examples/projects/dagster-slurm-example-hpc-workload/dagster_slurm_example_hpc_workload/ray/process_documents_docling_metaxy.py
@@ -12,7 +12,9 @@ docling workload (process_documents_docling.py).
 
 import glob
 import hashlib
+import math
 import os
+import re
 import time
 from pathlib import Path
 from typing import Any
@@ -31,6 +33,7 @@ import ray.data as rd
 from dagster_pipes import DagsterPipesError, PipesContext, open_dagster_pipes
 from dagster_slurm_example_shared import get_base_output_path
 from metaxy.ext.ray import MetaxyDatasink
+from ray.data import ActorPoolStrategy
 
 from dagster_slurm_example_hpc_workload.ray.process_documents_docling import (
     BasicDocumentConverter,
@@ -42,6 +45,8 @@ METAXY_TIMESTAMP_COLUMNS = (
     "metaxy_updated_at",
     "metaxy_deleted_at",
 )
+_PDF_PAGE_TOKEN_RE = re.compile(rb"/Type\s*/Page\b")
+_PDF_PAGES_TOKEN_RE = re.compile(rb"/Type\s*/Pages\b")
 
 
 def _coerce_metaxy_timestamps_to_utc(table: pa.Table) -> pa.Table:
@@ -69,6 +74,65 @@ def _upsert_column(table: pa.Table, name: str, array: pa.Array) -> pa.Table:
     return table.append_column(name, array)
 
 
+def _estimate_pdf_pages(path: str) -> int:
+    """Best-effort page estimate without external PDF dependencies."""
+    try:
+        data = Path(path).read_bytes()
+    except Exception:
+        return 0
+
+    page_tokens = len(_PDF_PAGE_TOKEN_RE.findall(data))
+    pages_tokens = len(_PDF_PAGES_TOKEN_RE.findall(data))
+    estimate = page_tokens - pages_tokens
+    return estimate if estimate > 0 else 0
+
+
+def _row_cost(row: dict[str, Any]) -> int:
+    page_estimate = _estimate_pdf_pages(str(row["source_path"]))
+    if page_estimate > 0:
+        return page_estimate
+
+    size_bytes = row.get("file_size_bytes")
+    if isinstance(size_bytes, (int, float)) and size_bytes > 0:
+        # Rough fallback: ~250KB ~= 1 page equivalent
+        return max(1, int(math.ceil(size_bytes / 250_000)))
+    return 1
+
+
+def _build_weighted_shards(
+    records: list[dict[str, Any]],
+    num_workers: int,
+) -> tuple[list[pa.Table], list[int]]:
+    target_shards = max(min(len(records), num_workers * 8), num_workers)
+    bins: list[dict[str, Any]] = [
+        {"cost": 0, "rows": []} for _ in range(max(target_shards, 1))
+    ]
+
+    weighted_records = [(record, _row_cost(record)) for record in records]
+    weighted_records.sort(key=lambda x: x[1], reverse=True)
+
+    for record, cost in weighted_records:
+        lightest = min(bins, key=lambda b: b["cost"])
+        lightest["rows"].append(record)
+        lightest["cost"] += cost
+
+    shard_tables: list[pa.Table] = []
+    shard_costs: list[int] = []
+    for bucket in bins:
+        rows = bucket["rows"]
+        if not rows:
+            continue
+        shard_tables.append(pl.DataFrame(rows).to_arrow())
+        shard_costs.append(int(bucket["cost"]))
+
+    return shard_tables, shard_costs
+
+
+def _sanitize_path_token(value: str) -> str:
+    token = re.sub(r"[^A-Za-z0-9._-]+", "_", value)
+    return token[:128] if token else "unknown"
+
+
 class MetaxyDoclingMapper:
     """Ray map_batches callable that converts documents and produces metaxy-compatible output."""
 
[…]
