+++ pyproject.toml
@@ -165,9 +165,15 @@ cwd = "slides"
 depends-on = ["slides-install"]
 description = "build static slides with base"
 
+[tool.pixi.feature.docs.tasks.slides-build-multimodal-base]
+cmd = "bun run build-multimodal-base"
+cwd = "slides"
+depends-on = ["slides-install"]
+description = "build static multimodal slides with base"
+
 [tool.pixi.feature.docs.tasks.docs-collect-slides]
-cmd = "bash -lc 'set -euo pipefail; rm -rf docs/static/slides; cp -R slides/dist/. docs/static/slides'"
-depends-on = ["slides-build-base"]
+cmd = "bash -lc 'set -euo pipefail; rm -rf docs/static/slides docs/static/slides-multimodal; cp -R slides/dist/. docs/static/slides; cp -R slides/dist-multimodal/. docs/static/slides-multimodal'"
+depends-on = ["slides-build-base", "slides-build-multimodal-base"]
 description = "Copy Slidev dist into Docusaurus static/"
 
 [tool.commitizen]

[…]
