### 2026-02-23 b3eac17
fix: generate data also on slurm
