+++ examples/projects/dagster-slurm-example/dagster_slurm_example/defs/ray/metaxy_ray_assets.py
@@ -6,17 +6,23 @@ Demonstrates metaxy incremental processing with Ray Data on HPC:
   using MetaxyDatasource/MetaxyDatasink for incremental reads/writes.
 
 In development mode, Ray runs locally. In production, it spawns a
-multi-node Ray cluster on Slurm. The HPC workload uses metaxy's
-DeltaLake store (METAXY_STORE=ray_embeddings) for distributed access.
+multi-node Ray cluster on Slurm. The HPC workload uses a deployment-
+specific metaxy store (dev vs prod) for distributed access.
 """
 
 import os
 
 import dagster as dg
-import metaxy as mx
 import metaxy.ext.dagster as mxd
-import polars as pl
-from dagster_slurm import ComputeResource, RayLauncher, SlurmRunConfig
+from dagster_slurm import BashLauncher, ComputeResource, RayLauncher, SlurmRunConfig
+
+
+def _ray_store_name_for_deployment(deployment: str) -> str:
+    return (
+        "ray_embeddings_prod"
+        if "supercomputer" in deployment or "production" in deployment
+        else "ray_embeddings_dev"
+    )
 
 
 @mxd.metaxify
@@ -24,44 +30,47 @@ from dagster_slurm import ComputeResource, RayLauncher, SlurmRunConfig
     metadata={"metaxy/feature": "ray_example/input_texts"},
     group_name="metaxy_ray",
 )
-def input_texts(store_ray: dg.ResourceParam[mx.MetadataStore]) -> pl.DataFrame:
-    """Register text samples in metaxy as a root feature.
+def input_texts(
+    context: dg.AssetExecutionContext,
+    compute_ray: ComputeResource,
+    config: SlurmRunConfig,
+):
+    """Register input_texts in the same runtime/store context as HPC workloads.
 
-    This lightweight asset runs on the dagster host (not HPC) and
-    registers the input data. The heavy embedding computation is
-    delegated to the downstream `compute_embeddings` asset.
+    Avoids host-local metaxy writes that may not be visible to remote Slurm jobs.
     """
-    samples = [
-        {"sample_uid": f"text_{i:03d}", "text": text}
-        for i, text in enumerate(
-            [
-                "The quick brown fox jumps over the lazy dog",
-                "Machine learning enables pattern recognition",
-                "High performance computing accelerates science",
-                "Distributed systems process data in parallel",
-                "Natural language processing understands text",
-                "Data pipelines transform raw inputs to features",
-                "Incremental processing saves compute resources",
-                "Metadata tracking ensures reproducibility",
-            ]
-        )
-    ]
-    df = pl.DataFrame(samples)
+    script_path = dg.file_relative_path(
+        __file__,
+        "../../../../dagster-slurm-example-hpc-workload/dagster_slurm_example_hpc_workload/ray/register_input_texts_metaxy.py",
+    )
+    metaxy_config = dg.file_relative_path(__file__, "../../../../../metaxy.toml")
+    deployment = os.getenv("DAGSTER_DEPLOYMENT", "development")
+    ray_store_name = _ray_store_name_for_deployment(deployment)
 
-    samples = df.with_columns(
-        pl.struct(
-            pl.col("text").cast(pl.String).alias("text"),
-        ).alias("metaxy_provenance_by_field")
+    context.log.info(
+        f"Registering ray_example/input_texts via launched workload.\n"
+        f"  Deployment: {deployment}\n"
+        f"  Store: {ray_store_name}"
     )
 
-    with store_ray:
-        increment = store_ray.resolve_update("ray_example/input_texts", samples=samples)
-        if len(increment.new) > 0:
[…]
