### 2026-02-23 6b59be0
fix: add missing dependency
