+++ docs/docusaurus.config.ts
@@ -72,6 +72,7 @@ const config: Config = {
       } satisfies Preset.Options,
     ],
   ],
+  plugins: ['docusaurus-lunr-search'],
 
   themeConfig: {
     // Replace with your project's social card
