+++ docs/docs/api/api_core.md
@@ -8,6 +8,14 @@ It is generated directly from the docstrings in the Python source code.
 This section covers the main components of the library.
 
 <a id="module-dagster_slurm"></a>
+<a id="dagster_slurm.ComputeLauncher"></a>
+<a id="dagster_slurm.SlurmRunConfig"></a>
+<a id="id0"></a>
+<a id="id22"></a>
+<a id="id30"></a>
+<a id="id48"></a>
+<a id="id63"></a>
+<a id="class-dagster_slurm-raylauncher"></a>
 
 Dagster Slurm Integration.
 
@@ -20,7 +28,7 @@ Run Dagster assets on Slurm clusters with support for:
 
 ### *class* dagster_slurm.BashLauncher(\*\*data)
 
-Bases: [`ComputeLauncher`](#dagster_slurm.ComputeLauncher)
+Bases: [`ComputeLauncher`](#class-dagster_slurmcomputelauncherdata)
 
 Executes Python scripts via bash.
 
@@ -139,7 +147,7 @@ slurm=slurm,
 
 #### default_extra_files *: `Optional`[`List`[`str`]]*
 
-#### default_launcher *: `Annotated`[[`ComputeLauncher`](#dagster_slurm.ComputeLauncher) | `PartialResource`]*
+#### default_launcher *: `Annotated`[[`ComputeLauncher`](#class-dagster_slurmcomputelauncherdata) | `PartialResource`]*
 
 #### default_skip_payload_upload *: `bool`*
 
@@ -151,7 +159,7 @@ Get appropriate Pipes client for this mode.
 
 - **Parameters:**
   - **context** (`InitResourceContext`) – Dagster resource context
-  - **launcher** (`Optional`[[`ComputeLauncher`](#dagster_slurm.ComputeLauncher)]) – Override launcher (uses default if None)
+  - **launcher** (`Optional`[[`ComputeLauncher`](#class-dagster_slurmcomputelauncherdata)]) – Override launcher (uses default if None)
 - **Returns:**
   LocalPipesClient or SlurmPipesClient
 
@@ -191,7 +199,7 @@ Execute asset with optional resource overrides.
 - **Parameters:**
   - **context** – Dagster execution context
   - **payload_path** (`str`) – Path to Python script
-  - **launcher** (`Optional`[[`ComputeLauncher`](#dagster_slurm.ComputeLauncher)]) – Override launcher for this asset
+  - **launcher** (`Optional`[[`ComputeLauncher`](#class-dagster_slurmcomputelauncherdata)]) – Override launcher for this asset
   - **extra_slurm_opts** (`Optional`[`Dict`[`str`, `Any`]]) – Override Slurm options (non-session mode)
     - nodes: int
     - cpus_per_task: int
@@ -210,7 +218,7 @@ Execute asset with optional resource overrides.
     or asset metadata key ‘skip_slurm_payload_upload’.
   - **remote_payload_path** (`Optional`[`str`]) – Remote path to an existing payload when skipping upload.
     Falls back to config.remote_payload_path or asset metadata.
-  - **config** (`Optional`[[`SlurmRunConfig`](#dagster_slurm.SlurmRunConfig)]) – Optional SlurmRunConfig for run-time configuration via launchpad.
+  - **config** (`Optional`[[`SlurmRunConfig`](#class-dagster_slurmslurmrunconfigconfig_dict)]) – Optional SlurmRunConfig for run-time configuration via launchpad.
     Values from config are used as defaults, but explicit parameters take precedence.
[…]
