### 2026-02-23 c561bbe
feat: webinar demonstration multimodal document processing (#101)
* feat: cancellation AND continued remote execution when dagster is restarted (#98)

* fix: refine cancellation vs automatic continuation for long running jobs where dagster is restarted

* fix: update deps for demo

* feat: add agster skills

* feat: first step towards skills here

* feat: metaxy basic version

* feat: mini ray example

* feat: docling example complete (for local execution)

* feat: slides first version

* fix: generate data also on slurm

* fix: add missing dependency

* fix: make it work somehow

* fix: stuff
[…]
