### 2026-02-23 4c37257
1.10.0-rc.3 [skip ci]
Automatically generated by python-semantic-release
