### 2026-02-23 7175b43
1.10.0-rc.3 [skip ci]
Automatically generated by python-semantic-release
