### 2026-02-23 69e1ebc
1.10.0-rc.3 [skip ci] (#103)
Automatically generated by python-semantic-release

Co-authored-by: semantic-release <semantic-release>
