+++ slides/slides_multimodal.md
@@ -94,20 +94,20 @@ class: bg-white text-black
 ---
 transition: slide-left
 layout: two-cols-header
+class: px-6
 ---
 
-# <span class="text-orange-500">The gap today</span>
+# <span class="text-orange-500 !text-5xl">Challenges</span>
 
 ::left::
 
-### HPC challenges
+### HPC pain points
 
 <v-clicks>
 
-- HPC users juggle Slurm scripts, modules, queues, and limited observability.
-- Packaging and reproducing Python/ML environments can be brittle.
-- Data teams want orchestration, lineage, retries, and fast local iteration.
-- Moving one asset from laptop prototype -> supercomputer often means rewrites.
+- Inconsistent Slurm scripts and limited obsevability create high operational overhead
+- Python/ML environments break inconsistent between environments laptop and cluster/s.
+- Moving one script from prototype to HPC often means rewriting runtime glue.
 
 </v-clicks>
 
@@ -121,10 +121,9 @@ layout: two-cols-header
 
 <v-clicks>
 
-- AI pipelines are expensive: unnecessary GPU/LLM recomputes burn time, budget, and cluster capacity.
-- Multimodal assets (docs, text, embeddings, media) change at different rates and need granular tracking.
-- Metaxy provides a topological feature container with field-level provenance across dependencies.
-- Operating principle: no unnecessary work
+- Coarse invalidation causes expensive GPU/LLM recomputes.
+- Multimodal fields change at different rates and need field-level tracking.
+- Metaxy resolves `new` and `stale` rows only, so unchanged data is skipped.
 
 </v-clicks>
 
@@ -155,9 +154,13 @@ transition: fade-out
 
 # Why orchestration matters
 
-- A raw “engine” (script, notebook, or binary) is not enough once you depend on sensors, ETL, or ML training.
+<div class="text-[20px] leading-relaxed space-y-4">
+
+- A raw engine is not enough once pipelines depend on ETL, sensors, or model training.
 - Orchestrators provide dependency tracking, retries, metrics, and the control plane HPC teams lack.
-- Dagster supplies that missing layer; dagster-slurm connects it to the supercomputer’s scheduler.
+- `dagster-slurm` maps that control plane directly to Slurm scheduling primitives.
+
+</div>
 
 ---
 transition: fade-out
@@ -260,19 +263,34 @@ image: /img/process_data_asset_view.png
 backgroundSize: contain
 ---
 
----
-transition: fade
----
+## transition: fadeclass: bg-slate-950 text-white
 
-# Backend of dagster-slurm
+# <span class="!text-5xl">Architecture</span>
 
-- **Dagster remains the control plane**: assets, schedules, and policies stay centralized.
-- **Reproducible packaging**: pixi/pixi-pack builds once and runs across local, edge, and HPC.
-- **Native Slurm execution**: dagster-slurm maps runs to `sbatch`/`squeue` with cluster constraints.
-- **Single pane of glass**: Slurm state, metrics, and Pipes logs flow back into Dagster UI.
+<div class="mt-8 w-full flex items-center justify-between gap-2 text-sm">
+  <span class="rounded-lg border border-slate-500 px-3 py-2">Dagster asset</span>
+  <span>-></span>
+  <span class="rounded-lg border border-emerald-400 px-3 py-2">dagster-slurm</span>
+  <span>-></span>
+  <span class="rounded-lg border border-slate-500 px-3 py-2">Slurm job</span>
+  <span>-></span>
[…]
