### 2026-02-23 #9795 @junghans
> > > **Paper file info**:
> > > ⚠️ Wordcount for `paper.md` is **1853**
> >
> >
> > [@geoHeil](https://github.com/geoHeil), this is nearly double the recommended max length (1000 words), can you try to shorten the paper a bit?
>
> [@junghans](https://github.com/junghans) https://joss.readthedocs.io/en/latest/paper.html states 1750 given it is now 1853 it is not double do you think that this current 1853 could still work? Should I try to shrink it to 1750? Cause half would really be tough..

I overlooked that update, the current length should be fine then.

Can you take care of the author tasks above?
