### 2026-02-23 #9795 @junghans
> - No DOI given, and none found for title: Ollama
> - No DOI given, and none found for title: LiteLLM
> - No DOI given, and none found for title: Open-WebUI
> - No DOI given, and none found for title: MCPs, Gatekeepers, and the Future of AI
> - No DOI given, and none found for title: Dagster: Data Orchestration Platform
> - No DOI given, and none found for title: Pixi: Reproducible Package Environments
> - No DOI given, and none found for title: Dagster Pipes Documentation
> - No DOI given, and none found for title: Existing Workflow Systems

I have confirmed those code, docs have no DOI.

@HPicatto maybe you could check zenodo.org another time as well.
