+++ examples/projects/dagster-slurm-example/dagster_slurm_example_tests/test_partitioned_metaxy_single_key.py
@@ -1,0 +1,274 @@
+from typing import Any, cast
+
+import dagster as dg
+import polars as pl
+from dagster_slurm_example.defs.metaxy import simple_metaxy_assets as assets
+
+
+class _FrameHandle:
+    def __init__(self, df: pl.DataFrame):
+        self._df = df
+
+    def to_polars(self) -> pl.DataFrame:
+        return self._df
+
+    def collect(self):
+        return self
+
+    def collect_schema(self):
+        return self._df.collect_schema()
+
+
+class _Increment:
+    def __init__(self, new_df: pl.DataFrame, stale_df: pl.DataFrame):
+        self.new = _FrameHandle(new_df)
+        self.stale = _FrameHandle(stale_df)
+
+
+class _FakeStore:
+    def __init__(
+        self,
+        increment: _Increment,
+        raw_numbers_df: pl.DataFrame,
+        processed_schema_df: pl.DataFrame,
+    ):
+        self._increment = increment
+        self._raw_numbers_df = raw_numbers_df
+        self._processed_schema_df = processed_schema_df
+        self.resolve_update_calls: list[dict[str, Any]] = []
+        self.writes: list[tuple[str, pl.DataFrame]] = []
+        self.deletes: list[dict[str, object]] = []
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, _exc_type, _exc_val, _exc_tb):
+        return False
+
+    def resolve_update(
+        self, feature_key: str, target_filters: list[Any] | None = None
+    ) -> _Increment:
+        self.resolve_update_calls.append(
+            {"feature_key": feature_key, "target_filters": target_filters}
+        )
+        return self._increment
+
+    def read(self, feature_key: str):
+        if feature_key == "example/partitioned_processed_numbers":
+            return _FrameHandle(self._processed_schema_df)
+        if feature_key == "example/raw_numbers":
+            return _FrameHandle(self._raw_numbers_df)
+        raise KeyError(feature_key)
+
+    def write(self, feature_key: str, frame: pl.DataFrame):
+        self.writes.append((feature_key, frame))
+
+    def delete(
+        self,
+        feature_key: str,
+        *,
+        filters,
+        soft: bool,
+        with_feature_history: bool,
+        with_sample_history: bool,
+    ):
+        self.deletes.append(
+            {
+                "feature_key": feature_key,
+                "filters": filters,
+                "soft": soft,
+                "with_feature_history": with_feature_history,
+                "with_sample_history": with_sample_history,
+            }
+        )
+
+
+def _call_partitioned_asset(
+    partition_key: str,
+    sample_uid: str | None,
+    subsample_pct: int | None,
+    store: _FakeStore,
+) -> pl.DataFrame:
+    partitioned_asset = cast(Any, assets.partitioned_processed_numbers)
+    fn = partitioned_asset.node_def.compute_fn.decorated_fn
+    context = dg.build_asset_context(partition_key=partition_key)
+    config = assets.PartitionedProcessedNumbersConfig(
+        sample_uid=sample_uid,
+        subsample_pct=subsample_pct,
+    )
+    return fn(context, store, config)
+
+
+def test_partitioned_single_key_forces_reprocess():
+    increment = _Increment(
+        new_df=pl.DataFrame(
+            {
+                "sample_uid": ["sample_001", "sample_999"],
+                "value": [10.0, 30.0],
+                "category": ["alpha", "beta"],
+            }
+        ),
+        stale_df=pl.DataFrame(schema={"sample_uid": pl.String, "value": pl.Float64}),
+    )
+    raw_numbers_df = pl.DataFrame(
+        {
[…]
