### 2026-02-25 ec91414
1.11.0 [skip ci]
Automatically generated by python-semantic-release
