### 2026-02-25 4d571d3
1.11.0 [skip ci]
Automatically generated by python-semantic-release
