### 2026-02-26 #9795 @junghans
@geoHeil small nitpick, while reading the paper I found the citation style bit misleading, in most cases hyperlinked "(Authors et al, YEAR)" would make it more readable than the currently used "Authors et al (YEAR)" with only the YEAR hyperlinked.

For a quick reference, the following citation commands can be used:
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"

I would use the second option throughout the paper, but as said this is just a nitpick.
