### 2026-02-27 a06ef50
1.11.0 [skip ci] (#107)
Automatically generated by python-semantic-release

Co-authored-by: semantic-release <semantic-release>
