### 2026-03-02 #9795 @junghans
> addressed in [ascii-supply-networks/dagster-slurm#107](https://github.com/ascii-supply-networks/dagster-slurm/pull/107)

This one looks a bit strange: "Eisty et al. (Eisty et al., 2025)", all others are good.
