### 2026-03-03 #9795 @danielskatz
You can use `@ref` in the .md file rather than `[@ref]` in this case
