### 2026-03-04 72d8589
paper.md: fix citation
