### 2026-03-04 #9795 @junghans
> should be fixed now

Not quite, it still is "Eisty et al. Eisty et al. (2025)", I fixed it here: https://github.com/ascii-supply-networks/dagster-slurm/pull/109
