### 2026-03-11 #9795 @junghans
> 1st one is done, 2nd partially release is in progress now [@HPicatto](https://github.com/HPicatto) can you take care of the zendodo stuff?

Any new on the release?
