### 2026-03-13 #9795 @junghans
@geoHeil can you work on this last item: "Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper." Authors are the same, but the title is not.
