### 2026-03-17 ascii-supply-networks/dagster-slurm#76 @geoHeil
@HPicatto I think we can almost close this one - can you just ahre the DOI for the final dcument here?
