### 2026-03-19 e2082dc
minor changes in paper.bib
