+++ docs/paper.bib
@@ -92,7 +92,7 @@
 
 @incollection{yoo2003slurm,
   author = {Yoo, Andy B. and Jette, Morris A. and Grondona, Mark},
-  title = {{SLURM}: Simple Linux Utility for Resource Management},
+  title = {{SLURM}: Simple {L}inux Utility for Resource Management},
   booktitle = {Job Scheduling Strategies for Parallel Processing},
   editor = {Feitelson, Dror G. and Rudolph, Larry and Schwiegelshohn, Uwe},
   series = {Lecture Notes in Computer Science},
@@ -182,7 +182,7 @@ abstract = {Reproducibility is widely acknowledged as a fundamental principle in
 
 @inproceedings{parsl,
   author    = {Babuji, Yadu and Woodard, Anna and Li, Zhuozhao and Katz, Daniel S. and Clifford, Ben and Kumar, Rohan and Lacinski, Lukasz and Chard, Ryan and Wozniak, Justin M. and Foster, Ian and Wilde, Michael and Chard, Kyle},
-  title     = {Parsl: Pervasive Parallel Programming in Python},
+  title     = {Parsl: Pervasive Parallel Programming in {P}ython},
   booktitle = {Proceedings of the 28th International Symposium on High-Performance Parallel and Distributed Computing},
   series    = {HPDC '19},
   pages     = {25--36},
@@ -193,7 +193,7 @@ abstract = {Reproducibility is widely acknowledged as a fundamental principle in
 
 @article{executorlib,
   author  = {Janssen, Jan and Taylor, Michael Gilbert and Yang, Ping and Neugebauer, Joerg and Perez, Danny},
-  title   = {executorlib -- Up-scaling Python workflows for hierarchical heterogeneous high-performance computing},
+  title   = {executorlib -- Up-scaling {P}ython workflows for hierarchical heterogeneous high-performance computing},
   journal = {Journal of Open Source Software},
   volume  = {10},
   number  = {108},
