### 2026-03-19 b346944
minor changes for JOSS publication (#110)
* minor changes in paper.md

* minor changes in paper.bib
