### 2026-03-19 ef16748
minor changes for JOSS publication (#110)
* minor changes in paper.md

* minor changes in paper.bib
