### 2026-03-19 #9795 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
