### 2026-03-19 #9795 @danielskatz
Congratulations to @HPicatto (Hernan Picatto) and co-authors on your JOSS publication!!

JOSS depends completely on volunteers - we couldn't have done this without you
